Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2       cat_3     cat_4
#> sample_1   -0.7910169  0.307595392  0.05628096 0.4271405
#> sample_2   -1.1714887  0.070224410  0.46825454 0.6330098
#> sample_3   -0.9176146 -0.087266254  0.18099773 0.8238831
#> sample_4   -0.8578756 -0.472213075  0.38644854 0.9436401
#> sample_5   -0.7021653 -0.427728443 -0.03568636 1.1655801
#> sample_6   -0.8380757 -0.021314606  0.10930558 0.7500848
#> sample_7   -0.9715665 -0.312320865  0.29067522 0.9932121
#> sample_8   -0.5253301 -0.608711720  0.26312725 0.8709146
#> sample_9   -0.8098275 -0.386970656  0.50841339 0.6883848
#> sample_10  -0.5699228 -0.059097195  0.17729158 0.4517284
#> sample_11  -1.2410288 -0.142416531  0.36840909 1.0150363
#> sample_12  -0.8161353 -0.149656347  0.13124604 0.8345456
#> sample_13  -1.0754729 -0.421546399  0.50271250 0.9943068
#> sample_14  -0.8762430  0.149609969  0.25685550 0.4697775
#> sample_15  -0.5369152 -0.536915195  0.21450089 0.8593295
#> sample_16  -0.7965207 -0.306972449  0.22933226 0.8741609
#> sample_17  -0.4497918  0.008041320  0.00804132 0.4337091
#> sample_18  -1.4553258  0.279275225  0.31774151 0.8583091
#> sample_19  -0.9249086 -0.044549917  0.17370365 0.7957549
#> sample_20  -0.4675094 -0.223887300  0.01534239 0.6760543
#> sample_21  -1.0084284 -0.105560653  0.21534707 0.8986420
#> sample_22  -0.4640351 -0.275982863  0.06048937 0.6795286
#> sample_23  -0.4285023 -0.307141465  0.28064520 0.4549986
#> sample_24  -0.9149728 -0.251678575  0.42620113 0.7404502
#> sample_25  -0.4802942 -0.408835214  0.03053145 0.8585979
#> sample_26  -1.0628392  0.079258212  0.49179049 0.4917905
#> sample_27  -0.7527864 -0.133747230  0.17298304 0.7135506
#> sample_28  -1.0445055 -0.390579029  0.28236544 1.1527191
#> sample_29  -0.9530999  0.256738062  0.06383440 0.6325274
#> sample_30  -0.7633880 -0.001247909  0.08390990 0.6807260
#> sample_31  -1.2992670  0.234663336  0.27312962 0.7914741
#> sample_32  -0.4916813 -0.420222381  0.10430209 0.8076016
#> sample_33  -1.2673701 -0.054347489  0.41227204 0.9094456
#> sample_34  -0.5463248 -0.231243751  0.20088960 0.5766789
#> sample_35  -1.0211035 -0.065592039  0.24113823 0.8455573
#> sample_36  -0.7690957 -0.100046072  0.26391930 0.6052225
#> sample_37  -1.1142462 -0.245208362  0.46393916 0.8955154
#> sample_38  -0.6186917 -0.458349012  0.21610604 0.8609346
#> sample_39  -0.5434462 -0.175721420  0.09254257 0.6266251
#> sample_40  -0.8737195 -0.210425258  0.09972967 0.9844151
#> sample_41  -1.4062420  0.403866642 -0.04811848 1.0504938
#> sample_42  -0.7123699  0.002283496  0.29527062 0.4148158
#> sample_43  -0.9512809  0.107326095  0.18579771 0.6581571
#> sample_44  -0.9706772 -0.459851531  0.55537915 0.8751495
#> sample_45  -0.7378343 -0.285849176  0.29518071 0.7285028
#> sample_46  -0.6855048 -0.210081090  0.32213572 0.5734502
#> sample_47  -0.8899540 -0.167819278  0.12698026 0.9307930
#> sample_48  -1.0731136  0.261887465  0.32971006 0.4815161
#> sample_49  -1.1333994  0.108313776  0.10831378 0.9167718
#> sample_50  -0.7956635 -0.184754409  0.06656002 0.9138579
#> sample_51  -0.6181208 -0.226078695  0.33739066 0.5068088
#> sample_52  -0.9131590 -0.191024330  0.29667888 0.8075045
#> sample_53  -1.0165891 -0.061077686  0.20718630 0.8704805
#> sample_54  -0.8480550 -0.128932344  0.35064074 0.6263466
#> sample_55  -0.8347674 -0.223858330  0.39518088 0.6634449
#> sample_56  -0.7521831 -0.477746255  0.23520355 0.9947258
#> sample_57  -0.7577189 -0.191323455  0.30863250 0.6404099
#> sample_58  -0.7128727 -0.390099299  0.28565614 0.8173158
#> sample_59  -0.8314881 -0.062354992  0.11589324 0.7779498
#> sample_60  -1.1502492  0.134949064  0.17662176 0.8386784
#> sample_61  -1.1706687  0.156202253  0.30743322 0.7070332
#> sample_62  -1.1329050  0.152293204  0.06334572 0.9172661
#> sample_63  -0.7470430  0.015097006 -0.03036537 0.7623114
#> sample_64  -0.8098275 -0.386970656  0.50841339 0.6883848
#> sample_65  -0.6855048 -0.210081090  0.32213572 0.5734502
#> sample_66  -1.2169486 -0.059495833  0.14635622 1.1300882
#> sample_67  -1.1218111 -0.190252879  0.42411342 0.8879505
#> sample_68  -0.8167585 -0.205849422  0.20909443 0.8135135
#> sample_69  -0.9052799 -0.183145231  0.23179862 0.8566266
#> sample_70  -0.6286940 -0.040907356  0.12272207 0.5468793
#> sample_71  -0.7668946 -0.147855345  0.33171774 0.5830322
#> sample_72  -0.5514304 -0.551430427  0.34056761 0.7622932
#> sample_73  -1.0578300  0.084267396  0.34499366 0.6285689
#> sample_74  -1.0588185  0.203423255  0.20342325 0.6519719
#> sample_75  -0.6713967 -0.195972975  0.16340103 0.7039686
#> sample_76  -0.6945341  0.020119259  0.06179196 0.6126229
#> sample_77  -0.2622003 -0.598672535  0.18978482 0.6710880
#> sample_78  -0.8524924 -0.083359313  0.31394248 0.6219092
#> sample_79  -0.5161673 -0.378016986  0.19532899 0.6988553
#> sample_80  -0.9393824 -0.011395638  0.23473743 0.7160406
#> sample_81  -0.9094702  0.018516593  0.01851659 0.8724370
#> sample_82  -0.8046081 -0.315059903  0.29400416 0.8256639
#> sample_83  -0.4908641 -0.194598308  0.14917323 0.5362892
#> sample_84  -0.9109118 -0.310137945  0.54633429 0.6747155
#> sample_85  -0.4138350 -0.645636638  0.32974301 0.7297287
#> sample_86  -0.5511553 -0.474194248  0.20026080 0.8250887
#> sample_87  -0.6774423 -0.099364455  0.07888378 0.6979230
#> sample_88  -0.5793231 -0.364211722  0.20913426 0.7344006
#> sample_89  -0.6087837 -0.448441067  0.14754237 0.9096824
#> sample_90  -0.6867232 -0.595751450  0.50286084 0.7796138
#> sample_91  -0.8955700 -0.232275773  0.24150858 0.8863372
#> sample_92  -0.5161673 -0.378016986  0.19532899 0.6988553
#> sample_93  -0.9006635 -0.178528827  0.19794874 0.8812436
#> sample_94  -0.5359044 -0.619285983  0.35609367 0.7990967
#> sample_95  -1.0358331 -0.030311194  0.29916801 0.7669762
#> sample_96  -0.7098601  0.086471337  0.12493762 0.4984511
#> sample_97  -0.7585256  0.003614481  0.04709959 0.7078115
#> sample_98  -1.0543044 -0.001154461  0.50032530 0.5551335
#> sample_99  -1.1550657  0.041185054  0.32303621 0.7908444
#> sample_100 -0.9127154 -0.249421223  0.39720594 0.7649307