Variance stabilizing transform from precision weights

Compute variance stabilizing transform (VST) from precision weights by scaling each observation by their respective weights

vst(x)

# S4 method for EList
vst(x)

Arguments

x

object storing data to be transformed

Details

A variance stabilizing transform is usually described in terms of a parametric model of the observed data. Instead, here inverse variance of each observation are stored in x$weight and the VST divides the observed data by the variances.

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# run crumblr on counts
cobj = crumblr(counts)

# run standard limma analysis on crumblr results
library(limma)

design = model.matrix(~Age, info)
fit = lmFit(vst(cobj), design)
fit = eBayes(fit)

topTable(fit, coef="Age", sort.by="none")
#>             logFC    AveExpr          t    P.Value adj.P.Val         B
#> cat_1 -0.02752732 -11.697014 -1.4576178 0.14809157 0.2961831 -4.688262
#> cat_2  0.05488603  -3.820013  0.8418307 0.40189937 0.4598290 -5.299368
#> cat_3  0.06462161   8.183378  0.7420073 0.45982903 0.4598290 -5.368209
#> cat_4 -0.15729771  30.013723 -1.6948122 0.09323535 0.2961831 -4.369430