Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1         cat_2        cat_3     cat_4
#> sample_1   -0.5399018 -0.2248207901  0.134553211 0.6301694
#> sample_2   -0.9933600 -0.0378485532  0.059789916 0.9714186
#> sample_3   -0.9216047  0.0063821057  0.095329592 0.8198930
#> sample_4   -0.9426017 -0.1122534261  0.486864805 0.5679903
#> sample_5   -0.1795238 -0.5472485545  0.127206493 0.5995658
#> sample_6   -0.5315218 -0.2164407332  0.064461652 0.6835009
#> sample_7   -0.5472486 -0.1795237744  0.127206493 0.5995658
#> sample_8   -0.5330673 -0.3179559349 -0.088381493 0.9394047
#> sample_9   -0.5992439 -0.2072018065  0.113705914 0.6927398
#> sample_10  -0.8510963 -0.1319736282  0.409157345 0.5739126
#> sample_11  -1.1289149 -0.0303026422  0.237961344 0.9212562
#> sample_12  -0.6968528 -0.1187749331  0.278526864 0.5371009
#> sample_13  -0.8560653 -0.0393041709  0.277033157 0.6183363
#> sample_14  -1.0903343 -0.1587761190  0.194863921 1.0542465
#> sample_15  -0.6881128  0.1082186120 -0.062406905 0.6423011
#> sample_16  -0.4530359  0.0047971540  0.044802489 0.4034363
#> sample_17  -1.2942538  0.0214229739  0.570530784 0.7023001
#> sample_18  -0.4960781 -0.5730391259  0.442191554 0.6269257
#> sample_19  -0.5687557 -0.1033924561  0.246281292 0.4258669
#> sample_20  -0.6573235 -0.1818997646  0.057329925 0.7818933
#> sample_21  -0.3403519 -0.4785022558 -0.060767055 0.8796212
#> sample_22  -0.7320793 -0.2212537161  0.155223855 0.7981092
#> sample_23  -0.8244205 -0.2135113889  0.274191817 0.7637400
#> sample_24  -1.0446204  0.0539919283  0.254662624 0.7359658
#> sample_25  -0.5346779 -0.3340071634  0.280359139 0.5883259
#> sample_26  -0.6142297 -0.1695438321  0.174227707 0.6095458
#> sample_27  -0.9618582  0.2122616114  0.175220340 0.5743763
#> sample_28  -0.8931296 -0.2923557121  0.316708351 0.8687769
#> sample_29  -0.8110359 -0.3214876808  0.355398979 0.7771246
#> sample_30  -0.9292059 -0.0488471377  0.207872709 0.7701803
#> sample_31  -1.1210263 -0.2519884388  0.548130861 0.8248839
#> sample_32  -1.2985260  0.1953990166  0.310911904 0.7922151
#> sample_33  -0.4166331 -0.3541127797  0.226917103 0.5438288
#> sample_34  -1.2695612  0.1392060195  0.182691131 0.9476640
#> sample_35  -0.9314963 -0.0035095307  0.167115986 0.7678898
#> sample_36  -0.7668946 -0.1478553451  0.331717735 0.5830322
#> sample_37  -0.7748172  0.1879935051 -0.012677190 0.5995009
#> sample_38  -1.4655305  0.2290652457  0.480379674 0.7560856
#> sample_39  -0.6234263 -0.0811019998  0.127989798 0.5765385
#> sample_40  -0.6385149 -0.0072431003  0.185660566 0.4600974
#> sample_41  -0.4208583  0.0769801311 -0.243927589 0.5878058
#> sample_42  -1.2415792 -0.0841263889  0.292351182 1.0333544
#> sample_43  -0.8536411  0.0520675186  0.133745550 0.6678280
#> sample_44  -0.8188780  0.0868305824 -0.099755373 0.8318028
#> sample_45  -1.1333994  0.1083137763  0.108313776 0.9167718
#> sample_46  -0.8601164 -0.0433553027  0.338579308 0.5648924
#> sample_47  -0.7330477 -0.3435829801  0.365564542 0.7110662
#> sample_48  -0.9158356  0.0121511658  0.057613540 0.8460709
#> sample_49  -0.4996353 -0.1533590799  0.176120121 0.4768743
#> sample_50  -1.0979540  0.0006583282  0.050668749 1.0466269
#> sample_51  -0.7092353 -0.1984096798  0.007442374 0.9002026
#> sample_52  -1.1203781  0.0758726309  0.075872631 0.9686329
#> sample_53  -0.9915647  0.1070476303 -0.088696947 0.9732140
#> sample_54  -0.5679696 -0.4195496029  0.220487752 0.7670315
#> sample_55  -1.1017783 -0.3708908107  0.650760437 0.8219087
#> sample_56  -0.5775446 -0.4291246422  0.314453392 0.6922159
#> sample_57  -1.0268426  0.0717696720  0.115254784 0.8398182
#> sample_58  -0.9259636 -0.0956152710  0.248156268 0.7734226
#> sample_59  -1.2756197 -0.0625970778  0.500872279 0.8373445
#> sample_60  -0.8046081 -0.3150599030  0.294004160 0.8256639
#> sample_61  -0.9584920  0.1401203036  0.215627856 0.6027438
#> sample_62  -1.0530566  0.0455556991  0.349767073 0.6577338
#> sample_63  -0.9095558  0.1890564478 -0.131851272 0.8523507
#> sample_64  -0.4315853 -0.2020108336  0.104719434 0.5288767
#> sample_65  -0.6794255 -0.2595716140  0.359467594 0.5795295
#> sample_66  -0.4201002 -0.3534088179 -0.073823956 0.8473330
#> sample_67  -0.4923841 -0.4923840862  0.219112233 0.7656559
#> sample_68  -0.8107173 -0.1442383399  0.094991349 0.8599643
#> sample_69  -0.6744931 -0.3134797716  0.480472108 0.5075008
#> sample_70  -0.6882420 -0.2987772055  0.028435706 0.9585835
#> sample_71  -1.2917219  0.1170453555  0.353434134 0.8212424
#> sample_72  -0.7974350 -0.2453663953  0.149946341 0.8928550
#> sample_73  -0.8680776 -0.3339950803  0.230534722 0.9715379
#> sample_74  -0.6667645 -0.2469106808  0.205074443 0.7086008
#> sample_75  -0.7564910 -0.1374517636  0.206319776 0.6876230
#> sample_76  -0.5813957 -0.3037639162  0.131554155 0.7536054
#> sample_77  -0.6535484 -0.2925350842  0.181249268 0.7648342
#> sample_78  -0.8710924 -0.0933878634 -0.040744130 1.0052244
#> sample_79  -0.9987027 -0.2102453047  0.300580319 0.9083677
#> sample_80  -0.9956235 -0.2696864704  0.373863769 0.8914462
#> sample_81  -0.9201158 -0.0397571051  0.138491126 0.8213818
#> sample_82  -0.9563038  0.3848701385  0.017145358 0.5542883
#> sample_83  -0.9556235  0.0613107156  0.288700558 0.6056123
#> sample_84  -0.4851002 -0.2970479484  0.290738716 0.4914094
#> sample_85  -0.6153834  0.0103225231 -0.316890388 0.9219512
#> sample_86  -1.4777029  0.3324057295  0.615531985 0.5297652
#> sample_87  -1.0204143 -0.2319568914  0.561994988 0.6903762
#> sample_88  -1.6109030 -0.0014650864  0.554060716 1.0583074
#> sample_89  -0.9140702 -0.1363656245  0.223008376 0.8274274
#> sample_90  -1.0497562  0.0033937029  0.385328314 0.6610342
#> sample_91  -0.6773902 -0.5030368067  0.248379282 0.9320477
#> sample_92  -0.9456921 -0.0653333921  0.395481811 0.6155437
#> sample_93  -0.5399018 -0.2248207901  0.134553211 0.6301694
#> sample_94  -0.8237404 -0.2716718477  0.374955317 0.7204570
#> sample_95  -0.6790018 -0.1509343902  0.155795877 0.6741403
#> sample_96  -0.9744202 -0.3866334960  0.428403502 0.9326502
#> sample_97  -0.7688686 -0.3460117433  0.136840028 0.9780403
#> sample_98  -0.6386125 -0.3401195213  0.157718907 0.8210131
#> sample_99  -0.5537985 -0.1360632968  0.120656550 0.5692052
#> sample_100 -0.5569610 -0.1392257514  0.154535367 0.5416513