Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -1.4228226 -0.123539643  0.723758218 0.8226041
#> sample_2   -1.1564607 -0.057848363  0.673039146 0.5412699
#> sample_3   -0.9072974  0.109636825 -0.076949131 0.8746097
#> sample_4   -1.1334580 -0.087489466  0.344643889 0.8763036
#> sample_5   -1.1599143  0.206961934  0.166956599 0.7859958
#> sample_6   -0.9414877  0.031961460  0.195590884 0.7139353
#> sample_7   -0.4639232 -0.204412031 -0.263252531 0.9315878
#> sample_8   -0.6689527 -0.307939379  0.338687786 0.6382043
#> sample_9   -1.1414104 -0.042798145  0.336691477 0.8475171
#> sample_10  -0.8380757 -0.021314606  0.109305577 0.7500848
#> sample_11  -0.7564910 -0.137451764  0.206319776 0.6876230
#> sample_12  -1.1825308  0.363393725  0.222811774 0.5963253
#> sample_13  -0.4250929 -0.303732079  0.218457303 0.5103677
#> sample_14  -1.1182054 -0.019593093  0.166992863 0.9708056
#> sample_15  -1.0026320 -0.214174659  0.332369048 0.8844376
#> sample_16  -0.7965207 -0.306972449  0.229332260 0.8741609
#> sample_17  -0.7193005 -0.396527076  0.347050958 0.7687766
#> sample_18  -1.0269888 -0.179690944  0.499067499 0.7076123
#> sample_19  -0.6373225  0.114093549  0.035621934 0.4876071
#> sample_20  -0.6197624  0.093187425 -0.077438092 0.6040130
#> sample_21  -0.6142297 -0.169543832  0.174227707 0.6095458
#> sample_22  -0.7617131 -0.487276260  0.301181100 0.9478083
#> sample_23  -0.8434226 -0.124299901  0.289675897 0.6780466
#> sample_24  -0.8493218  0.056386790  0.098059486 0.6948756
#> sample_25  -0.5660750 -0.350963590  0.106869504 0.8101691
#> sample_26  -0.9277617  0.045687408  0.089172520 0.7929018
#> sample_27  -0.6591369 -0.360643862  0.348503660 0.6712771
#> sample_28  -0.6567537 -0.358260709  0.318625951 0.6963885
#> sample_29  -1.4577050  0.106270559  0.688875865 0.6625586
#> sample_30  -1.0409158  0.221325886  0.057696462 0.7618935
#> sample_31  -1.2509311 -0.093478315  0.358506808 0.9859026
#> sample_32  -0.3418917 -0.227481346 -0.124827192 0.6942002
#> sample_33  -0.5322007 -0.532200743  0.180749065 0.8836524
#> sample_34  -0.4923841 -0.492384086  0.219112233 0.7656559
#> sample_35  -1.0190904 -0.013568537  0.164679694 0.8679792
#> sample_36  -1.0554269 -0.401500404  0.349915685 1.1070116
#> sample_37  -0.6332069 -0.472864244  0.342172754 0.7638984
#> sample_38  -0.6299487  0.001323048  0.083001079 0.5456246
#> sample_39  -0.7864246 -0.511987719  0.586624570 0.7117877
#> sample_40  -0.7518917 -0.241066059  0.346720606 0.6462371
#> sample_41  -0.9450659  0.229053901  0.028383205 0.6876288
#> sample_42  -0.7715096 -0.009369535  0.154259889 0.6266192
#> sample_43  -0.6149034 -0.222861276  0.277094675 0.5606700
#> sample_44  -0.9268143  0.131792636  0.001172453 0.7938492
#> sample_45  -0.8208103 -0.209901218  0.242083906 0.7886276
#> sample_46  -0.9100852 -0.246791004  0.367575298 0.7893009
#> sample_47  -0.9933600 -0.037848553  0.059789916 0.9714186
#> sample_48  -0.7625458 -0.143506550  0.270469248 0.6355831
#> sample_49  -0.6829347 -0.360161330  0.079205330 0.9638907
#> sample_50  -0.9325239 -0.102175555  0.311800243 0.7228992
#> sample_51  -1.3420175 -0.105254846  0.174330016 1.2729423
#> sample_52  -0.7330477 -0.343582980  0.365564542 0.7110662
#> sample_53  -0.4654535 -0.398762172  0.278124488 0.5860912
#> sample_54  -1.1131882  0.083062522  0.035434473 0.9946912
#> sample_55  -1.0060719 -0.050560463  0.136025492 0.9206069
#> sample_56  -1.1554152 -0.056802883  0.513741976 0.6984761
#> sample_57  -0.9277617  0.089172520  0.045687408 0.7929018
#> sample_58  -0.8113087 -0.259240115  0.251585509 0.8189633
#> sample_59  -1.0186677 -0.171369843  0.384155959 0.8058816
#> sample_60  -0.5757136 -0.021402845  0.205986997 0.3911294
#> sample_61  -0.8799610 -0.216666774  0.136973266 0.9596545
#> sample_62  -1.2851191 -0.019452768  0.521678205 0.7828937
#> sample_63  -0.9481812  0.025267900  0.261656678 0.6612567
#> sample_64  -0.9802314 -0.320985742  0.354769696 0.9464474
#> sample_65  -1.1401287 -0.149729949  0.529028494 0.7608301
#> sample_66  -1.4545848  0.240010906  0.355523793 0.8590501
#> sample_67  -0.8931296 -0.292355712  0.316708351 0.8687769
#> sample_68  -1.0746826 -0.504137733  0.729815904 0.8490044
#> sample_69  -0.7157931 -0.204967433  0.046346995 0.8744135
#> sample_70  -0.6829347 -0.360161330  0.079205330 0.9638907
#> sample_71  -0.8622937  0.125092920  0.125092920 0.6121079
#> sample_72  -1.6470170  0.257220474  0.450124140 0.9396724
#> sample_73  -0.7496407 -0.238815059  0.316710744 0.6717450
#> sample_74  -0.4916813 -0.420222381  0.104302087 0.8076016
#> sample_75  -1.0120228 -0.223565434  0.423061731 0.8125265
#> sample_76  -0.6825801 -0.154512645  0.189258894 0.6478338
#> sample_77  -0.7406126 -0.288627497  0.325738806 0.7035013
#> sample_78  -0.8650714 -0.002847915  0.333624322 0.5342950
#> sample_79  -1.0251553 -0.069643835  0.274127704 0.8206714
#> sample_80  -0.6142297 -0.169543832  0.174227707 0.6095458
#> sample_81  -0.6332069 -0.472864244  0.342172754 0.7638984
#> sample_82  -0.7071355 -0.035967181  0.268244193 0.4748584
#> sample_83  -1.0092769 -0.220819568  0.393546735 0.8365498
#> sample_84  -1.1608755  0.080837619  0.317226397 0.7628115
#> sample_85  -0.8900235 -0.427399938  0.646211048 0.6712124
#> sample_86  -0.4866327 -0.243010629  0.189122727 0.5405206
#> sample_87  -0.8588094  0.003414154  0.239802932 0.6155923
#> sample_88  -0.6044879 -0.109791683  0.026783852 0.6874958
#> sample_89  -0.4640351 -0.275982863  0.060489374 0.6795286
#> sample_90  -0.8017269 -0.190817846  0.103981694 0.8885631
#> sample_91  -0.2462409 -0.757066512  0.461090928 0.5422165
#> sample_92  -0.7556725 -0.086622844  0.131630722 0.7106646
#> sample_93  -0.7439229 -0.233097257  0.254605949 0.7224142
#> sample_94  -0.7387132 -0.119674020  0.066911936 0.7914753
#> sample_95  -0.6071088 -0.270636528  0.284889275 0.5928560
#> sample_96  -1.1383230  0.057927710  0.188547893 0.8918474
#> sample_97  -0.7062694  0.050056728  0.128528344 0.5276843
#> sample_98  -1.1432222  0.053028595  0.223654112 0.8665395
#> sample_99  -0.7635350 -0.211466440 -0.044412355 1.0194138
#> sample_100 -1.0251553 -0.069643835  0.274127704 0.8206714