Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2       cat_3     cat_4
#> sample_1   -1.1449835  0.003639191  0.29740031 0.8439440
#> sample_2   -0.9024653  0.025521487 -0.02210656 0.8990504
#> sample_3   -0.5288283 -0.018002634 -0.16110348 0.7079344
#> sample_4   -0.9920021 -0.266065093  0.34299897 0.9150682
#> sample_5   -0.7340812 -0.065031551 -0.01740350 0.8165162
#> sample_6   -0.3407258 -0.089411417  0.03575173 0.3943855
#> sample_7   -1.1476302  0.048620546  0.25771234 0.8412973
#> sample_8   -0.8467029 -0.029941729  0.17915007 0.6974945
#> sample_9   -0.8071506 -0.255082015  0.21870234 0.8435303
#> sample_10  -0.8136661 -0.324117900  0.38502962 0.7527544
#> sample_11  -1.1392705  0.102442665  0.14592778 0.8909000
#> sample_12  -0.7996824 -0.376825561  0.36675247 0.8097555
#> sample_13  -0.7067551 -0.455440649  0.35959635 0.8025994
#> sample_14  -0.5928919 -0.315260195  0.23128351 0.6768686
#> sample_15  -0.7737807 -0.162871661 -0.05465808 0.9913105
#> sample_16  -0.6847410 -0.013572680  0.02991243 0.6684012
#> sample_17  -1.5852980 -0.235371269  0.86324102 0.9574282
#> sample_18  -0.5946776 -0.202635470  0.07826692 0.7190461
#> sample_19  -0.4039127 -0.635714336  0.23612463 0.8035024
#> sample_20  -0.5895787 -0.197536645  0.04169304 0.7454223
#> sample_21  -1.1154626 -0.069494067  0.21140832 0.9735484
#> sample_22  -0.8998301 -0.236535924  0.27428970 0.8620764
#> sample_23  -1.1253706 -0.079402012  0.27997199 0.9248006
#> sample_24  -0.9774852 -0.389698545  0.45759932 0.9095844
#> sample_25  -1.2857913  0.077513578  0.35936473 0.8489130
#> sample_26  -0.9374269  0.161185389  0.03602225 0.7402193
#> sample_27  -0.6332069 -0.472864244  0.34217275 0.7638984
#> sample_28  -0.5907864 -0.254314192  0.12216338 0.7229372
#> sample_29  -0.8122442 -0.201335069  0.17514250 0.8384367
#> sample_30  -0.7051115  0.009541888  0.16672747 0.5288421
#> sample_31  -1.0120228 -0.223565434  0.42306173 0.8125265
#> sample_32  -1.0349286 -0.132060894  0.49112770 0.6758618
#> sample_33  -0.7359705  0.111327391 -0.16957499 0.7942181
#> sample_34  -0.4640351 -0.275982863  0.06048937 0.6795286
#> sample_35  -0.5857113 -0.141025495 -0.04338703 0.7701238
#> sample_36  -0.4684806 -0.280428412  0.09604916 0.6528599
#> sample_37  -0.7956635 -0.184754409  0.06656002 0.9138579
#> sample_38  -0.9149728 -0.251678575  0.42620113 0.7404502
#> sample_39  -0.6108392  0.020432624 -0.06851486 0.6589214
#> sample_40  -1.0446204  0.053991928  0.25466262 0.7359658
#> sample_41  -0.4111326 -0.348612240  0.16221338 0.5975315
#> sample_42  -0.5858834 -0.370772006  0.27277823 0.6838772
#> sample_43  -0.8998301 -0.236535924  0.27428970 0.8620764
#> sample_44  -0.7564910 -0.137451764  0.20631978 0.6876230
#> sample_45  -0.8440562 -0.074923084  0.21883804 0.7001412
#> sample_46  -0.7407633 -0.174367806  0.14653991 0.7685912
#> sample_47  -0.8071506 -0.255082015  0.21870234 0.8435303
#> sample_48  -1.2941393  0.069165557  0.45110017 0.7738736
#> sample_49  -1.0359741 -0.133106374  0.51865559 0.6504249
#> sample_50  -0.8210344 -0.154555463  0.16635226 0.8092376
#> sample_51  -0.3987622 -0.465453547  0.27812449 0.5860912
#> sample_52  -1.4580068  0.519155872  0.15143109 0.7874199
#> sample_53  -0.8497703 -0.130647616  0.38017801 0.6002399
#> sample_54  -1.3759018 -0.354250504  0.77067909 0.9594732
#> sample_55  -0.9181220 -0.140417421  0.25599785 0.8025416
#> sample_56  -0.9006635 -0.178528827  0.19794874 0.8812436
#> sample_57  -0.8212310 -0.269162416  0.34520389 0.7451895
#> sample_58  -1.1364367 -0.146037987  0.47300122 0.8094735
#> sample_59  -0.9195970 -0.256302786  0.50986953 0.6660303
#> sample_60  -0.4377777 -0.437777716  0.12675209 0.7488033
#> sample_61  -0.5652357 -0.488274649  0.32676235 0.7267480
#> sample_62  -0.7870512 -0.512614384  0.61231521 0.6873504
#> sample_63  -0.4334705 -0.504929478  0.20656684 0.7318331
#> sample_64  -0.4626340  0.035204381  0.18091619 0.2465135
#> sample_65  -0.4768007 -0.347588931  0.29903823 0.5253514
#> sample_66  -1.1432222  0.053028595  0.22365411 0.8665395
#> sample_67  -0.9364049 -0.056046195  0.27343301 0.7190181
#> sample_68  -0.6165227 -0.028736019  0.01474909 0.6305096
#> sample_69  -0.6881128 -0.062406905  0.10821861 0.6423011
#> sample_70  -0.7051115  0.009541888  0.16672747 0.5288421
#> sample_71  -0.5294891 -0.328818392  0.21772531 0.6405822
#> sample_72  -0.6893936 -0.213969949  0.46557158 0.4377920
#> sample_73  -0.4325510 -0.059875708 -0.05987571 0.5523024
#> sample_74  -0.6471787 -0.348685670  0.22466031 0.7712040
#> sample_75  -0.6220318 -0.323538861  0.04913642 0.8964343
#> sample_76  -0.5032545 -0.156978234  0.24032356 0.4199091
#> sample_77  -1.0316202 -0.377693730  0.21009294 1.1992210
#> sample_78  -0.6856954 -0.157627950  0.22186167 0.6214617
#> sample_79  -0.9149728 -0.251678575  0.42620113 0.7404502
#> sample_80  -0.7727643  0.032860870  0.11453890 0.6253645
#> sample_81  -0.5179655 -0.258454294  0.03634525 0.7400745
#> sample_82  -0.8254424 -0.158963512  0.20041049 0.7839955
#> sample_83  -0.5582123 -0.558212272  0.42760452 0.6888200
#> sample_84  -1.2591076  0.006558761  0.27482275 0.9777261
#> sample_85  -1.1549312  0.130267080  0.21194511 0.8127190
#> sample_86  -0.4877017 -0.191435853  0.11529441 0.5638431
#> sample_87  -0.5182533 -0.446794313  0.36029713 0.6047505
#> sample_88  -1.0042058 -0.278268838  0.46213123 0.8203435
#> sample_89  -0.9943445 -0.205887105  0.26789725 0.9323343
#> sample_90  -1.1097884 -0.307441966  0.55936871 0.8578617
#> sample_91  -1.1447734 -0.098804815  0.46466454 0.7789136
#> sample_92  -1.0349286 -0.132060894  0.49112770 0.6758618
#> sample_93  -0.9072974 -0.076949131  0.10963682 0.8746097
#> sample_94  -0.7404284 -0.229602733  0.22238239 0.7476487
#> sample_95  -0.7364874 -0.225661765  0.18928209 0.7728671
#> sample_96  -0.7577189 -0.191323455  0.30863250 0.6404099
#> sample_97  -1.2892029 -0.023536575  0.63038989 0.6823496
#> sample_98  -0.9942445 -0.146946666  0.18952557 0.9516656
#> sample_99  -0.4908641 -0.194598308  0.14917323 0.5362892
#> sample_100 -0.8833722 -0.420748715  0.48810704 0.8160139