Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts
count data with samples as rows and variables are columns

pseudocount
added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also