Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -0.9732512 -0.247314192  0.210518902 1.0100465
#> sample_2   -0.6992423  0.017435378 -0.247257176 0.9290641
#> sample_3   -0.5410539 -0.281542683  0.240646699 0.5819499
#> sample_4   -0.6794255 -0.008257186 -0.008257186 0.6959398
#> sample_5   -0.6825478 -0.591575988  0.447317066 0.8268067
#> sample_6   -1.3024967  0.149755603  0.410481865 0.7422593
#> sample_7   -0.7511061 -0.082056507  0.096191724 0.7369709
#> sample_8   -0.3043611 -0.248791233 -0.053046656 0.6061990
#> sample_9   -0.8710924 -0.093387863 -0.040744130 1.0052244
#> sample_10  -1.0898459 -0.435919463  0.714652564 0.8111128
#> sample_11  -0.5712869 -0.293655148  0.059984892 0.8049571
#> sample_12  -0.9358514 -0.158146834  0.493615130 0.6003831
#> sample_13  -0.3418917 -0.227481346 -0.124827192 0.6942002
#> sample_14  -1.0370484 -0.234701967  0.061563849 1.2101866
#> sample_15  -0.7854816 -0.511044766  0.560538850 0.7359875
#> sample_16  -0.6386125 -0.340119521  0.157718907 0.8210131
#> sample_17  -0.7918246 -0.239756062  0.113883979 0.9176967
#> sample_18  -0.9356779 -0.007691104  0.201400694 0.7419683
#> sample_19  -0.5544023 -0.405982287  0.118542181 0.8418424
#> sample_20  -0.9512809  0.185797710  0.107326095 0.6581571
#> sample_21  -1.0832113 -0.429284841  0.585945838 0.9265503
#> sample_22  -0.8001936 -0.448795739  0.599172816 0.6498165
#> sample_23  -1.0446204  0.254662624  0.053991928 0.7359658
#> sample_24  -0.7537522 -0.242926533  0.376112676 0.6205660
#> sample_25  -1.4378926  0.126082902  0.372215971 0.9395938
#> sample_26  -1.1434383 -0.153039612  0.635417749 0.6610602
#> sample_27  -1.2906399  0.203285124  0.243290459 0.8440643
#> sample_28  -0.7128727 -0.390099299  0.285656139 0.8173158
#> sample_29  -1.2577715 -0.159159187  0.518720521 0.8982101
#> sample_30  -0.5422471 -0.076883845 -0.031421471 0.6505524
#> sample_31  -0.6667017 -0.305688379  0.308677924 0.6637122
#> sample_32  -0.8376467 -0.226737613  0.452020830 0.6123635
#> sample_33  -0.4430375 -0.163452596  0.166026605 0.4404635
#> sample_34  -0.7746702 -0.285121934  0.087553351 0.9722387
#> sample_35  -0.6442413 -0.283227920  0.112084817 0.8153844
#> sample_36  -1.0498096  0.092287825  0.249473409 0.7080483
#> sample_37  -0.6794255 -0.008257186 -0.008257186 0.6959398
#> sample_38  -1.3912825 -0.221211219  0.585880221 1.0266135
#> sample_39  -0.7449947 -0.293009570  0.384870139 0.6531341
#> sample_40  -1.1515712  0.044679578  0.290812647 0.8160790
#> sample_41  -0.8583015 -0.041540413  0.308133336 0.5917086
#> sample_42  -0.9320791 -0.154374533  0.409094824 0.6773588
#> sample_43  -0.6892685 -0.111190614  0.182570504 0.6178886
#> sample_44  -0.6515544 -0.176130704  0.019613873 0.8080712
#> sample_45  -1.0041525 -0.156854612  0.258089240 0.9029178
#> sample_46  -1.2667081  0.048968694  0.267222260 0.9505171
#> sample_47  -1.2917561  0.071548710  0.421222459 0.7989850
#> sample_48  -0.4997631 -0.105858812  0.103232986 0.5023889
#> sample_49  -0.7696231 -0.150583885  0.390547088 0.5296599
#> sample_50  -0.3742951 -0.109602578 -0.064140204 0.5480379
#> sample_51  -0.8390232  0.066685461  0.023200349 0.7491374
#> sample_52  -0.9848288 -0.081961115  0.068321088 0.9984689
#> sample_53  -0.6267791 -0.466436466  0.280777936 0.8124376
#> sample_54  -0.8622937  0.125092920  0.125092920 0.6121079
#> sample_55  -0.4965805  0.139408306 -0.425121497 0.7822937
#> sample_56  -1.2619469 -0.163334650  0.574264293 0.8510173
#> sample_57  -0.2707550 -0.392115864 -0.064902953 0.7277738
#> sample_58  -1.2855411  0.166711252  0.248389283 0.8704405
#> sample_59  -0.5753821 -0.360270754  0.176033955 0.7596189
#> sample_60  -0.8328779 -0.166399004  0.265734351 0.7335426
#> sample_61  -0.5992439 -0.207201807  0.113705914 0.6927398
#> sample_62  -1.0290265 -0.181728653  0.553382727 0.6573724
#> sample_63  -0.8329769 -0.016215781  0.072731705 0.7764610
#> sample_64  -0.7346478 -0.282662629  0.263881078 0.7534293
#> sample_65  -0.9300802 -0.152375613  0.379841201 0.7026146
#> sample_66  -0.9036574 -0.240363226  0.306180480 0.8378402
#> sample_67  -0.8130069  0.049216582 -0.093884262 0.8576746
#> sample_68  -0.8141439  0.002617196 -0.045010853 0.8565376
#> sample_69  -0.6539754 -0.355482388  0.288067851 0.7213899
#> sample_70  -0.7987454 -0.413082968  0.031602854 1.1802256
#> sample_71  -0.5544023 -0.405982287  0.118542181 0.8418424
#> sample_72  -0.9024653  0.025521487 -0.022106562 0.8990504
#> sample_73  -0.9497860 -0.021799220  0.360135391 0.6114498
#> sample_74  -0.4880279 -0.094123585 -0.005176099 0.5873276
#> sample_75  -0.5930054 -0.050681109 -0.098309158 0.7419957
#> sample_76  -1.4275148  0.418311932  0.038822311 0.9703805
#> sample_77  -0.7658348 -0.491397940  0.332777503 0.9244552
#> sample_78  -0.4915320 -0.123807244 -0.232020829 0.8473601
#> sample_79  -0.7020454 -0.191219789 -0.032995783 0.9262610
#> sample_80  -1.2851191 -0.019452768  0.521678205 0.7828937
#> sample_81  -0.7891081 -0.366251298  0.273786057 0.8815734
#> sample_82  -0.6099982 -0.217956153  0.214177202 0.6137772
#> sample_83  -0.3181997 -0.595831412  0.262830207 0.6512009
#> sample_84  -1.1313127 -0.140913982  0.414611820 0.8576148
#> sample_85  -0.9292504 -0.207115674  0.555024378 0.5813417
#> sample_86  -0.9994380 -0.152140159  0.224337412 0.9272408
#> sample_87  -0.8548426 -0.392219065  0.243769701 1.0032920
#> sample_88  -0.9549013  0.018547887  0.355020123 0.5813332
#> sample_89  -1.0530566  0.045555699  0.349767073 0.6577338
#> sample_90  -0.6591369 -0.360643862  0.348503660 0.6712771
#> sample_91  -1.1368710 -0.090902489  0.375717043 0.8520565
#> sample_92  -0.8856620 -0.222367827  0.172944910 0.9350850
#> sample_93  -0.9671745 -0.379387805  0.367826597 0.9787357
#> sample_94  -0.6068829 -0.214840848  0.181574425 0.6401494
#> sample_95  -0.8329769 -0.016215781  0.072731705 0.7764610
#> sample_96  -1.1688721  0.116326187  0.343716029 0.7088298
#> sample_97  -0.5715798 -0.423159774  0.252595664 0.7421439
#> sample_98  -0.8212310 -0.269162416  0.345203886 0.7451895
#> sample_99  -1.4108741  0.002819224  0.362193225 1.0458617
#> sample_100 -1.0681304  0.155644997  0.398206634 0.5142788