Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -1.0686258  0.193615890  0.334197841 0.5408121
#> sample_2   -0.6818879 -0.103810005  0.114443561 0.6712543
#> sample_3   -0.8208103 -0.209901218  0.242083906 0.7886276
#> sample_4   -1.1447126 -0.046100339  0.367875459 0.8229375
#> sample_5   -0.8760808  0.111305829  0.322614923 0.4421601
#> sample_6   -0.9530864  0.020362777  0.324574151 0.6081494
#> sample_7   -1.1079493 -0.238911426  0.404638813 0.9422219
#> sample_8   -0.9407658 -0.110417545  0.430713428 0.6204700
#> sample_9   -0.4782032 -0.348991459  0.328888249 0.4983064
#> sample_10  -0.7496407 -0.238815059  0.316710744 0.6717450
#> sample_11  -1.0589128  0.039699488  0.523496439 0.4957169
#> sample_12  -0.7095009 -0.458186515  0.389111346 0.7785761
#> sample_13  -0.9328606  0.125746392  0.040588583 0.7665256
#> sample_14  -0.7097447 -0.038576468  0.388867547 0.3594537
#> sample_15  -0.6586420 -0.238788139  0.137689432 0.7597407
#> sample_16  -0.8426421 -0.025880943  0.144744574 0.7237784
#> sample_17  -0.7058908 -0.253905648  0.056249280 0.9035471
#> sample_18  -1.1660647  0.119133515  0.312037181 0.7348940
#> sample_19  -0.5946776 -0.202635470  0.078266916 0.7190461
#> sample_20  -1.4249663  0.184471576  0.227956687 1.0125381
#> sample_21  -0.5195411 -0.381390723  0.227673341 0.6732584
#> sample_22  -1.3783198 -0.279707559  0.598842845 1.0591846
#> sample_23  -1.1376894 -0.039077146  0.304694393 0.8720722
#> sample_24  -0.7206160 -0.331151247  0.242194734 0.8095725
#> sample_25  -0.4493749  0.048463495 -0.033214536 0.4341260
#> sample_26  -0.9127154 -0.082367138  0.145891514 0.8491911
#> sample_27  -0.3868982 -0.219844091  0.227468127 0.3792741
#> sample_28  -1.0388988 -0.033376954  0.330588424 0.7416873
#> sample_29  -0.6924154 -0.164347964  0.315225116 0.5415382
#> sample_30  -0.8484519  0.013771622  0.138934765 0.6957455
#> sample_31  -1.0122399 -0.164942066  0.322761141 0.8544209
#> sample_32  -0.6794255 -0.008257186 -0.008257186 0.6959398
#> sample_33  -0.8694513 -0.483788845  0.502027949 0.8512122
#> sample_34  -0.7839488 -0.783948765  0.493711756 1.0741858
#> sample_35  -0.9013283 -0.070979971  0.072120873 0.9001874
#> sample_36  -1.2776403 -0.064617631  0.529157076 0.8131008
#> sample_37  -0.6218140 -0.177128151  0.270184067 0.5287581
#> sample_38  -0.8007808 -0.311232600  0.262113380 0.8499000
#> sample_39  -1.1618199 -0.013197204  0.667679884 0.5073372
#> sample_40  -0.4031374 -0.049497368  0.177892475 0.2747423
#> sample_41  -0.7844855 -0.022345420  0.345379360 0.4614515
#> sample_42  -0.8289868 -0.109864141  0.158399845 0.7804511
#> sample_43  -0.7875436  0.018081550  0.311068675 0.4583934
#> sample_44  -1.4144185  0.051918595  0.320182581 1.0423173
#> sample_45  -0.6068829 -0.214840848  0.181574425 0.6401494
#> sample_46  -0.8863267 -0.108622139  0.041660064 0.9532888
#> sample_47  -0.5579063 -0.003595568 -0.003595568 0.5650974
#> sample_48  -0.6949778 -0.305513037  0.067162248 0.9333286
#> sample_49  -1.6387210  0.176568966  0.492906294 0.9692457
#> sample_50  -0.8403516 -0.071218550  0.185501297 0.7260689
#> sample_51  -0.8459425 -0.126819790  0.320492428 0.6522698
#> sample_52  -0.5276307 -0.268119455  0.108358116 0.6873920
#> sample_53  -1.2680822 -0.002415842  0.341355697 0.9291424
#> sample_54  -0.6033047 -0.211262593  0.148111408 0.6664559
#> sample_55  -0.4138350 -0.645636638  0.329743010 0.7297287
#> sample_56  -0.3855090 -0.218454877  0.195520921 0.4084429
#> sample_57  -0.6299487  0.001323048  0.083001079 0.5456246
#> sample_58  -0.9715665 -0.312320865  0.290675219 0.9932121
#> sample_59  -0.6336113 -0.335118354  0.122714739 0.8460150
#> sample_60  -0.5986018 -0.689573616  0.556363388 0.7318121
#> sample_61  -0.5322964 -0.331625721  0.249404162 0.6145180
#> sample_62  -0.3950876 -0.466546533 -0.048811333 0.9104454
#> sample_63  -0.8452189 -0.459556416  0.291859672 1.0129156
#> sample_64  -0.8837998 -0.161665121  0.089649307 0.9558157
#> sample_65  -0.6071088 -0.270636528  0.284889275 0.5928560
#> sample_66  -0.7956635 -0.184754409  0.066560020 0.9138579
#> sample_67  -0.5983513 -0.320719565  0.293646737 0.6254241
#> sample_68  -0.9899937 -0.402207016  0.671403971 0.7207967
#> sample_69  -0.8380757 -0.021314606  0.109305577 0.7500848
#> sample_70  -1.4108741  0.002819224  0.362193225 1.0458617
#> sample_71  -0.4981374 -0.672490745  0.214812450 0.9558157
#> sample_72  -1.2451505 -0.146538211  0.400005495 0.9916832
#> sample_73  -1.2085766 -0.387596014  0.567915431 1.0282571
#> sample_74  -0.8289868 -0.109864141  0.158399845 0.7804511
#> sample_75  -0.9942445 -0.146946666  0.189525570 0.9516656
#> sample_76  -0.9374269  0.161185389  0.036022247 0.7402193
#> sample_77  -0.6794255 -0.259571614  0.359467594 0.5795295
#> sample_78  -0.8084984 -0.385641528  0.481169146 0.7129708
#> sample_79  -1.1545586 -0.005935848  0.391365950 0.7691285
#> sample_80  -1.0565868  0.042025536  0.409750317 0.6048109
#> sample_81  -0.7346478 -0.282662629  0.263881078 0.7534293
#> sample_82  -0.8452888 -0.178809900  0.444378692 0.5797200
#> sample_83  -0.6336113 -0.335118354  0.122714739 0.8460150
#> sample_84  -1.1168508 -0.247812976  0.492587089 0.8720767
#> sample_85  -0.6997412 -0.121663350  0.339151853 0.4822527
#> sample_86  -0.5631453 -0.171103180 -0.118459447 0.8527079
#> sample_87  -1.0172029 -0.114335157  0.282080116 0.8494579
#> sample_88  -1.1540144 -0.055402074  0.485728900 0.7236875
#> sample_89  -0.2933096 -0.481361793  0.228879821 0.5457915
#> sample_90  -1.2638221  0.001844309  0.308574576 0.9534032
#> sample_91  -0.7383418 -0.348876984  0.451242316 0.6359764
#> sample_92  -0.8244205 -0.213511389  0.274191817 0.7637400
#> sample_93  -0.5986725 -0.262200299  0.189784825 0.6710880
#> sample_94  -0.8622860 -0.560005115  0.723341277 0.6989498
#> sample_95  -1.0371135 -0.081602027  0.397971053 0.7207444
#> sample_96  -0.9094380 -0.187303331  0.264681793 0.8320596
#> sample_97  -0.7677072  0.119595963 -0.005567180 0.6536784
#> sample_98  -0.7638411  0.041784099  0.041784099 0.6802729
#> sample_99  -1.0881524 -0.285805972  0.354231383 1.0197270
#> sample_100 -0.8463871 -0.179908199  0.499633330 0.5266620