Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1         cat_2        cat_3     cat_4
#> sample_1   -0.8848420 -2.840682e-01  0.252236551 0.9166736
#> sample_2   -1.5846704 -1.632847e-01  0.643806711 1.1041484
#> sample_3   -0.7460073 -7.695768e-02  0.059617852 0.7633471
#> sample_4   -0.9220226 -9.167430e-02  0.215055965 0.7986409
#> sample_5   -0.5115132 -2.866141e-02  0.128524171 0.4116504
#> sample_6   -0.6213140 -2.014601e-01 -0.093246543 0.9160206
#> sample_7   -0.8219016 -3.323533e-01  0.603107315 0.5511476
#> sample_8   -0.7408860 -2.420830e-02 -0.024208296 0.7893026
#> sample_9   -0.8327685 -2.218594e-01  0.365927255 0.6887006
#> sample_10  -0.7310051 -6.452614e-02 -0.308148224 1.1036794
#> sample_11  -0.8619090  8.547235e-02  0.163943964 0.6124927
#> sample_12  -0.4337672 -1.901451e-01 -0.190145130 0.8140575
#> sample_13  -0.8122442 -2.013351e-01  0.175142502 0.8384367
#> sample_14  -1.1131882  3.543447e-02  0.083062522 0.9946912
#> sample_15  -1.1409317  7.690988e-03  0.264410834 0.8688299
#> sample_16  -0.5475045 -2.879934e-01  0.361351204 0.4741467
#> sample_17  -0.9024653 -2.210656e-02  0.025521487 0.8990504
#> sample_18  -1.0573904  1.263797e-01  0.277610674 0.6534000
#> sample_19  -0.9119001 -3.111263e-01  0.624334397 0.5986920
#> sample_20  -1.2482131  1.745331e-02  0.204039263 1.0267205
#> sample_21  -0.6574894 -2.964761e-01  0.214349571 0.7396159
#> sample_22  -0.7651867 -5.741315e-01  0.495066943 0.8442512
#> sample_23  -0.6165227  1.474909e-02 -0.028736019 0.6305096
#> sample_24  -0.5762666  9.818845e-02  0.059722169 0.4183560
#> sample_25  -1.2517518 -2.156599e-01  0.584459386 0.8829524
#> sample_26  -0.7817378 -1.260468e-02 -0.170828686 0.9651711
#> sample_27  -1.1501212 -5.150888e-02  0.428064201 0.7735658
#> sample_28  -0.7668946 -1.478553e-01  0.331717735 0.5830322
#> sample_29  -0.5472486 -1.795238e-01  0.127206493 0.5995658
#> sample_30  -0.5262416 -3.255709e-01  0.185254690 0.6665579
#> sample_31  -1.2816330 -6.861040e-02  0.610931133 0.7393123
#> sample_32  -0.7488858 -1.824903e-01  0.213924925 0.7174512
#> sample_33  -0.9127154  1.458915e-01 -0.082367138 0.8491911
#> sample_34  -0.6021791 -3.245474e-01  0.353332317 0.5733942
#> sample_35  -1.2855411  1.667113e-01  0.248389283 0.8704405
#> sample_36  -0.8150297 -2.629611e-01  0.283582593 0.7944082
#> sample_37  -0.6035051 -3.258734e-01  0.382311655 0.5470669
#> sample_38  -0.8594569  2.090187e-02 -0.196162636 1.0347176
#> sample_39  -1.0130448 -1.101771e-01  0.249196944 0.8740249
#> sample_40  -0.6573235  5.732992e-02 -0.181899765 0.7818933
#> sample_41  -1.0210736  7.753873e-02  0.077538732 0.8659961
#> sample_42  -0.8401479 -4.544854e-01  0.258464382 1.0361690
#> sample_43  -0.6894915 -5.151381e-01  0.343523555 0.8611060
#> sample_44  -1.0588185  2.034233e-01  0.203423255 0.6519719
#> sample_45  -0.6301921 -4.698495e-01  0.311851090 0.7881905
#> sample_46  -0.7119542 -2.599691e-01  0.093670954 0.8782523
#> sample_47  -0.7011320 -3.116672e-01  0.104493203 0.9083060
#> sample_48  -0.5477010 -6.310826e-01  0.627872327 0.5509113
#> sample_49  -1.1675534  1.593176e-01  0.274830445 0.7334054
#> sample_50  -0.7340812 -1.740350e-02 -0.065031551 0.8165162
#> sample_51  -0.1795238 -5.472486e-01  0.127206493 0.5995658
#> sample_52  -0.6941726 -7.513336e-02 -0.183346945 0.9526529
#> sample_53  -0.7918062 -3.022580e-01  0.195580432 0.8984838
#> sample_54  -0.8390232  2.320035e-02  0.066685461 0.7491374
#> sample_55  -0.9527890 -5.253450e-01  0.543853398 0.9342806
#> sample_56  -0.6708132 -3.097999e-01  0.368079856 0.6125332
#> sample_57  -1.2798783  1.288889e-01  0.254052039 0.8969374
#> sample_58  -1.6169733  1.006782e-01  0.444449693 1.0718455
#> sample_59  -0.7092353 -1.984097e-01  0.007442374 0.9002026
#> sample_60  -0.4883323 -4.883323e-01  0.186122758 0.7905418
#> sample_61  -0.7954262  2.709252e-01  0.130343234 0.3941578
#> sample_62  -0.4630721 -3.963807e-01  0.247169510 0.6122833
#> sample_63  -1.1450395  1.401587e-01  0.140158717 0.8647221
#> sample_64  -0.5227820 -1.550572e-01 -0.057418764 0.7352580
#> sample_65  -0.7157931  4.634700e-02 -0.204967433 0.8744135
#> sample_66  -0.7696231 -1.505839e-01  0.390547088 0.5296599
#> sample_67  -1.0077032  9.090905e-02 -0.002181370 0.9189756
#> sample_68  -1.0371135 -8.160203e-02  0.397971053 0.7207444
#> sample_69  -0.8244205 -2.135114e-01  0.274191817 0.7637400
#> sample_70  -0.6899302  2.472322e-02  0.024723217 0.6404837
#> sample_71  -1.1502446 -1.042761e-01  0.575265475 0.6792552
#> sample_72  -0.6161604 -5.327788e-01  0.325882828 0.8230564
#> sample_73  -0.4638684 -6.996416e-02 -0.167602625 0.7014352
#> sample_74  -0.6142297 -1.695438e-01  0.174227707 0.6095458
#> sample_75  -0.5623366 -4.853755e-01  0.296325037 0.7513871
#> sample_76  -1.1334580 -8.748947e-02  0.344643889 0.8763036
#> sample_77  -0.6701383 -2.502844e-01  0.237418789 0.6830039
#> sample_78  -1.1607835  3.546725e-02  0.385141000 0.7401753
#> sample_79  -0.5488862 -4.004662e-01  0.082385541 0.8669669
#> sample_80  -1.1553480  1.715229e-01  0.171522936 0.8123021
#> sample_81  -0.8080212 -3.184729e-01  0.325077314 0.8014168
#> sample_82  -0.9374269 -1.070786e-01  0.372494483 0.6720110
#> sample_83  -1.1383230  5.792771e-02  0.188547893 0.8918474
#> sample_84  -0.9374269 -1.070786e-01  0.372494483 0.6720110
#> sample_85  -0.7956635 -1.847544e-01  0.066560020 0.9138579
#> sample_86  -0.6328541 -2.718408e-01  0.038314176 0.8663807
#> sample_87  -0.9384847 -4.276590e-01  0.285290763 1.0808529
#> sample_88  -1.1485937  2.902042e-05  0.329508222 0.8190564
#> sample_89  -1.4310878  8.525965e-02  0.379020765 0.9668074
#> sample_90  -0.8689607 -2.681868e-01  0.147973594 0.9891739
#> sample_91  -0.9201158 -3.975711e-02  0.138491126 0.8213818
#> sample_92  -1.0599141  1.238560e-01  0.309573169 0.6264849
#> sample_93  -0.7364874 -2.256618e-01  0.189282087 0.7728671
#> sample_94  -0.6207796 -1.260834e-01  0.167677764 0.5791852
#> sample_95  -0.6699268 -5.789550e-01  0.328602040 0.9202798
#> sample_96  -0.7511061 -8.205651e-02  0.096191724 0.7369709
#> sample_97  -1.0358331 -3.031119e-02  0.299168007 0.7669762
#> sample_98  -1.6202156  1.474463e-01  0.404166134 1.0686032
#> sample_99  -1.2941393  6.916556e-02  0.451100168 0.7738736
#> sample_100 -0.8975997 -3.635172e-01  0.503293493 0.7578234