Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -0.5597515 -0.559751528  0.455479152 0.6640239
#> sample_2   -1.2298032 -0.260402659  0.483175375 1.0070305
#> sample_3   -0.5322964 -0.331625721  0.249404162 0.6145180
#> sample_4   -1.0835219 -0.281175416  0.321820668 1.0428766
#> sample_5   -0.7807485  0.066549347  0.145020963 0.5691782
#> sample_6   -0.6703747  0.085951385 -0.142307267 0.7267306
#> sample_7   -0.6481461 -0.416344456  0.359494441 0.7049961
#> sample_8   -0.5197318 -0.448272788  0.389124002 0.5788805
#> sample_9   -0.7147228  0.041603234  0.260656800 0.4124628
#> sample_10  -1.0323385 -0.026816670  0.266944449 0.7922108
#> sample_11  -1.0478678  0.050744470  0.287133248 0.7099901
#> sample_12  -0.9786228 -0.252685764  0.245152665 0.9861559
#> sample_13  -0.6985074 -0.120429541  0.309133119 0.5098038
#> sample_14  -0.5900211 -0.429678484  0.035684766 0.9840149
#> sample_15  -1.0320523  0.151717828  0.066560020 0.8137744
#> sample_16  -0.4916813 -0.420222381  0.104302087 0.8076016
#> sample_17  -0.9356779 -0.007691104  0.201400694 0.7419683
#> sample_18  -0.7585256  0.003614481  0.047099593 0.7078115
#> sample_19  -0.5596401 -0.141904947  0.187574255 0.5139708
#> sample_20  -0.7276216 -0.404848177  0.461962497 0.6705072
#> sample_21  -1.0657054  0.118064658  0.430936979 0.5167038
#> sample_22  -0.5225371 -0.321866400  0.151917952 0.6924855
#> sample_23  -1.2713245  0.044352291  0.301072138 0.9259001
#> sample_24  -0.6207796 -0.126083354  0.167677764 0.5791852
#> sample_25  -0.9164612 -0.194326524  0.327862858 0.7829249
#> sample_26  -1.0415509 -0.036028996  0.361272802 0.7163071
#> sample_27  -1.1035201 -0.057551498  0.138193079 1.0228785
#> sample_28  -1.2630977  0.189154607  0.100207121 0.9737360
#> sample_29  -1.0809601 -0.427033588  0.558783206 0.9492104
#> sample_30  -1.0358331 -0.030311194  0.299168007 0.7669762
#> sample_31  -1.6452932  0.122368729  0.671476540 0.8514479
#> sample_32  -0.6848048 -0.593832976  0.475365428 0.8032723
#> sample_33  -0.6818879 -0.103810005  0.114443561 0.6712543
#> sample_34  -0.5294891 -0.328818392  0.217725314 0.6405822
#> sample_35  -0.7801729 -0.428775041  0.318439360 0.8905086
#> sample_36  -0.8167585 -0.205849422  0.209094430 0.8135135
#> sample_37  -0.9597914  0.098815563  0.284532709 0.5764431
#> sample_38  -0.6125788 -0.276106565  0.402651878 0.4860335
#> sample_39  -0.8385957 -0.172116806  0.327839146 0.6828734
#> sample_40  -1.0214026  0.614352663 -0.232945198 0.6399951
#> sample_41  -0.6108392  0.020432624 -0.068514862 0.6589214
#> sample_42  -0.7269858 -0.275000661  0.198783691 0.8032028
#> sample_43  -0.7564910 -0.137451764  0.206319776 0.6876230
#> sample_44  -0.7076304  0.007022975  0.199926641 0.5006808
#> sample_45  -0.6667645 -0.246910681  0.205074443 0.7086008
#> sample_46  -0.9070705 -0.243776249  0.337253634 0.8135931
#> sample_47  -0.6794255 -0.008257186 -0.008257186 0.6959398
#> sample_48  -0.8955700 -0.232275773  0.241508580 0.8863372
#> sample_49  -1.1120512 -0.013438936  0.129661908 0.9958283
#> sample_50  -1.2798783  0.254052039  0.128888896 0.8969374
#> sample_51  -1.0544135  0.244869450  0.129356563 0.6801875
#> sample_52  -0.9418643 -0.061505566  0.335796231 0.6675736
#> sample_53  -0.9356779 -0.007691104  0.201400694 0.7419683
#> sample_54  -0.9149728 -0.251678575  0.426201134 0.7404502
#> sample_55  -0.6361647 -0.004892907  0.152292677 0.4887649
#> sample_56  -0.9094380 -0.187303331  0.264681793 0.8320596
#> sample_57  -0.9216047  0.095329592  0.006382106 0.8198930
#> sample_58  -0.6713967 -0.195972975  0.163401026 0.7039686
#> sample_59  -0.6891762 -0.366402846  0.116448926 0.9391302
#> sample_60  -0.9497860 -0.021799220  0.360135391 0.6114498
#> sample_61  -0.9374269 -0.107078597  0.372494483 0.6720110
#> sample_62  -0.6574894 -0.296476052  0.214349571 0.7396159
#> sample_63  -1.6095452 -0.058947766  0.649237292 1.0192557
#> sample_64  -1.0544135  0.129356563  0.244869450 0.6801875
#> sample_65  -0.6629392 -0.243085360  0.171858493 0.7341661
#> sample_66  -0.7930375 -0.370180652  0.305574786 0.8576434
#> sample_67  -1.0147081  0.083904159  0.038441785 0.8923622
#> sample_68  -0.7487670 -0.296781871  0.576158420 0.4693904
#> sample_69  -1.0456623 -0.040140470  0.420674734 0.6651281
#> sample_70  -0.6055237 -0.063199369 -0.017736994 0.6864600
#> sample_71  -1.1549312  0.130267080  0.211945111 0.8127190
#> sample_72  -1.0962069 -0.164648713  0.230664023 1.0301916
#> sample_73  -0.5983513 -0.320719565  0.293646737 0.6254241
#> sample_74  -1.2756197 -0.062597078  0.500872279 0.8373445
#> sample_75  -1.0478678  0.050744470  0.287133248 0.7099901
#> sample_76  -0.8908555 -0.227561320  0.207756751 0.9106601
#> sample_77  -0.4722905 -0.746727347  0.200653972 1.0183639
#> sample_78  -0.6794255 -0.259571614  0.359467594 0.5795295
#> sample_79  -0.8385397 -0.227630631  0.479701185 0.5864692
#> sample_80  -1.1493792  0.092333897  0.217497040 0.8395483
#> sample_81  -0.8868783 -0.424254748  0.542588263 0.7685448
#> sample_82  -1.3024967  0.149755603  0.410481865 0.7422593
#> sample_83  -1.2277421 -0.329800539  0.608469099 0.9490736
#> sample_84  -1.0235553 -0.176257414  0.442781794 0.7570309
#> sample_85  -1.2879143  0.027762458  0.457325117 0.8028268
#> sample_86  -0.9220226 -0.091674303  0.215055965 0.7986409
#> sample_87  -0.4363770  0.021456105 -0.109164078 0.5240850
#> sample_88  -0.8212310 -0.269162416  0.345203886 0.7451895
#> sample_89  -1.1207384 -0.130339719  0.321645405 0.9294327
#> sample_90  -0.9325239 -0.102175555  0.311800243 0.7228992
#> sample_91  -1.2693036 -0.111850804  0.537493754 0.8436606
#> sample_92  -1.0419917 -0.086480230  0.484064628 0.6444073
#> sample_93  -0.5465389 -0.469577844  0.166410923 0.8497058
#> sample_94  -0.6432477  0.069702112  0.145209664 0.4283359
#> sample_95  -0.5481300 -0.333018669 -0.005805758 0.8869545
#> sample_96  -0.8480550 -0.128932344  0.350640737 0.6263466
#> sample_97  -1.2941393  0.069165557  0.451100168 0.7738736
#> sample_98  -0.5949680 -0.258495765  0.156448087 0.6970157
#> sample_99  -0.7597333 -0.090683630  0.166036217 0.6843807
#> sample_100 -1.0349852  0.018164685  0.227256483 0.7895641