Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

Usage

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1         cat_2        cat_3     cat_4
#> sample_1   -1.2809675 -1.530114e-02  0.464271938 0.8319967
#> sample_2   -1.2936636  1.585887e-01  0.315774294 0.8193006
#> sample_3   -1.0306880  2.246191e-02  0.193087422 0.8151387
#> sample_4   -1.1004989 -1.101002e-01  0.184699376 1.0258997
#> sample_5   -0.9801961 -3.924095e-01  0.486140941 0.8864646
#> sample_6   -1.0096061 -2.836691e-01  0.545023527 0.7482518
#> sample_7   -0.7628189 -5.717636e-01  0.467129409 0.8674531
#> sample_8   -0.8607511  4.495751e-02  0.202143092 0.6136505
#> sample_9   -0.6125788 -2.761066e-01  0.402651878 0.4860335
#> sample_10  -0.9407658 -1.104175e-01  0.430713428 0.6204700
#> sample_11  -1.3142320  2.581647e-01  0.398746638 0.6573206
#> sample_12  -0.8955643 -1.734296e-01  0.163042625 0.9059513
#> sample_13  -0.6357143 -4.039127e-01  0.236124633 0.8035024
#> sample_14  -0.9774852 -3.896985e-01  0.457599315 0.9095844
#> sample_15  -1.1218111 -1.902529e-01  0.424113424 0.8879505
#> sample_16  -0.6562751 -5.653033e-01  0.231028091 0.9905503
#> sample_17  -0.9580956  1.535359e-02  0.471370980 0.4713710
#> sample_18  -1.4693878  3.764389e-01  0.340720776 0.7522282
#> sample_19  -1.2948215  1.991036e-01  0.277575166 0.8181428
#> sample_20  -0.9220226 -9.167430e-02  0.215055965 0.7986409
#> sample_21  -1.2798783  1.288889e-01  0.254052039 0.8969374
#> sample_22  -1.0287655 -7.325401e-02  0.306235616 0.7957838
#> sample_23  -0.7597384 -1.406992e-01  0.238790400 0.6616473
#> sample_24  -1.1685749  2.767586e-02  0.556935184 0.5839639
#> sample_25  -0.6174719 -1.227757e-01  0.133944148 0.6063035
#> sample_26  -0.5271130 -3.889626e-01  0.320184875 0.5958908
#> sample_27  -0.6755976 -2.557438e-01  0.299782015 0.6315594
#> sample_28  -0.9216047  6.382106e-03  0.095329592 0.8198930
#> sample_29  -0.6924154 -1.643480e-01  0.315225116 0.5415382
#> sample_30  -0.2276720 -6.197141e-01  0.395516584 0.4518695
#> sample_31  -0.9732512 -2.473142e-01  0.210518902 1.0100465
#> sample_32  -0.6612466 -4.868932e-01  0.144378585 1.0037612
#> sample_33  -0.8929752 -1.152706e-01  0.080473966 0.9277718
#> sample_34  -0.7761455 -4.247476e-01  0.286748681 0.9141445
#> sample_35  -1.2977619  6.554296e-02  0.509035465 0.7231835
#> sample_36  -0.8210344 -1.545555e-01  0.166352257 0.8092376
#> sample_37  -0.7466184 -2.946332e-01  0.413551813 0.6276998
#> sample_38  -0.5157275  4.880232e-02  0.087268600 0.3796566
#> sample_39  -1.1207384 -1.303397e-01  0.321645405 0.9294327
#> sample_40  -0.9176146  1.809977e-01 -0.087266254 0.8238831
#> sample_41  -0.9330312 -5.267246e-02  0.241088660 0.7446150
#> sample_42  -1.2792036  3.647319e-02  0.365952393 0.8767780
#> sample_43  -0.6229518 -4.626092e-01  0.248887156 0.8366738
#> sample_44  -0.9356779 -7.691104e-03  0.201400694 0.7419683
#> sample_45  -0.5533879 -1.856631e-01  0.193826534 0.5452244
#> sample_46  -1.1449835  3.639191e-03  0.297400310 0.8439440
#> sample_47  -0.6936642 -2.249591e-02  0.102667234 0.6134929
#> sample_48  -0.5974203  3.385146e-02 -0.152734495 0.7163034
#> sample_49  -0.9848288 -8.196112e-02  0.068321088 0.9984689
#> sample_50  -0.8289868 -1.098641e-01  0.158399845 0.7804511
#> sample_51  -1.3768774 -3.552261e-01  0.795345912 0.9367576
#> sample_52  -1.0084126 -1.611148e-01  0.290870361 0.8786570
#> sample_53  -0.7597333 -9.068363e-02  0.166036217 0.6843807
#> sample_54  -1.4507595  2.438362e-01  0.322307843 0.8846154
#> sample_55  -0.7723812 -5.570349e-02  0.226147661 0.6019370
#> sample_56  -0.6328541 -2.718408e-01  0.038314176 0.8663807
#> sample_57  -1.0438019 -3.828000e-02  0.391282664 0.6907992
#> sample_58  -1.1139610 -3.116145e-01  0.638577759 0.7869978
#> sample_59  -0.6427375 -4.109359e-01  0.299305698 0.7543677
#> sample_60  -0.8799610 -2.166668e-01  0.136973266 0.9596545
#> sample_61  -1.1004989 -1.101002e-01  0.184699376 1.0258997
#> sample_62  -0.9052799 -1.831452e-01  0.231798621 0.8566266
#> sample_63  -0.7358833 -5.448280e-01  0.251503375 1.0292079
#> sample_64  -0.9292504 -2.071157e-01  0.581341687 0.5550244
#> sample_65  -1.1253706 -7.940201e-02  0.279971988 0.9248006
#> sample_66  -0.4251215 -4.965805e-01  0.139408306 0.7822937
#> sample_67  -1.2635920  1.451752e-01  0.145175180 0.9732417
#> sample_68  -0.5015121  6.301768e-02 -0.062145461 0.5006399
#> sample_69  -0.7330477 -3.435830e-01  0.365564542 0.7110662
#> sample_70  -0.5789065 -1.342207e-01 -0.084210280 0.7973375
#> sample_71  -0.8241940  1.231873e-01 -0.105071329 0.8060780
#> sample_72  -0.6777102 -2.578563e-01  0.329930323 0.6056362
#> sample_73  -0.6790018 -1.509344e-01  0.155795877 0.6741403
#> sample_74  -0.8966489 -2.958750e-01  0.347675201 0.8448487
#> sample_75  -1.3041328  1.046345e-01  0.532078475 0.6674198
#> sample_76  -0.5751257 -2.386535e-01  0.012660933 0.8011183
#> sample_77  -0.7288123 -4.060389e-01  0.489345189 0.6455059
#> sample_78  -0.6611358 -3.626428e-01  0.377757283 0.6460213
#> sample_79  -0.6761292 -1.653035e-01 -0.165303533 1.0067362
#> sample_80  -1.2671623 -1.097095e-01  0.509329746 0.8675420
#> sample_81  -0.9147999 -3.444115e-02  0.102134386 0.8471066
#> sample_82  -0.7776489  2.208799e-01 -0.015508848 0.5722778
#> sample_83  -1.1482069 -1.022383e-01  0.520950247 0.7294950
#> sample_84  -0.8210344  1.663523e-01 -0.154555463 0.8092376
#> sample_85  -0.4613941 -4.613941e-01  0.433989964 0.4887982
#> sample_86  -1.2488699 -1.502576e-01  0.430772245 0.9683553
#> sample_87  -0.3341754 -4.633871e-01  0.139608995 0.6579535
#> sample_88  -0.8583015 -4.154041e-02  0.308133336 0.5917086
#> sample_89  -0.4756451 -8.613075e-01  0.416352977 0.9205996
#> sample_90  -0.8460216 -1.795426e-01  0.553344874 0.4722193
#> sample_91  -0.3862720 -7.090454e-01  0.253765359 0.8415520
#> sample_92  -1.2695612  1.392060e-01  0.182691131 0.9476640
#> sample_93  -0.8931296 -2.923557e-01  0.316708351 0.8687769
#> sample_94  -0.7963157 -4.449178e-01  0.493351825 0.7478817
#> sample_95  -1.2882273  1.205399e-01  0.321210575 0.8464769
#> sample_96  -0.9314963  1.671160e-01 -0.003509531 0.7678898
#> sample_97  -0.9349584 -1.572538e-01  0.465934776 0.6262774
#> sample_98  -1.1793248  1.875515e-01  0.366599687 0.6251737
#> sample_99  -1.0211035 -6.559204e-02  0.241138229 0.8455573
#> sample_100 -1.1485937  2.902042e-05  0.329508222 0.8190564