Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2       cat_3     cat_4
#> sample_1   -0.3533413 -0.676114708  0.04056297 0.9888931
#> sample_2   -1.0366687 -0.133800939  0.54574059 0.6247290
#> sample_3   -0.4295295 -0.500988509  0.17346654 0.7570515
#> sample_4   -0.9597914  0.284532709  0.09881556 0.5764431
#> sample_5   -0.5654158 -0.287784037  0.02237089 0.8308289
#> sample_6   -0.5606570 -0.345545582  0.07061481 0.8355877
#> sample_7   -0.7224714 -0.270486309  0.16483176 0.8281260
#> sample_8   -1.1549312  0.211945111  0.13026708 0.8127190
#> sample_9   -0.7564910 -0.137451764  0.20631978 0.6876230
#> sample_10  -0.5150647 -0.515064681  0.07272198 0.9574074
#> sample_11  -0.6847410 -0.013572680  0.02991243 0.6684012
#> sample_12  -0.4280666 -0.198492127  0.06977186 0.5567868
#> sample_13  -1.1398858 -0.093917244  0.40603871 0.8277643
#> sample_14  -0.8631123 -0.046351144  0.42655524 0.4829082
#> sample_15  -0.9294581 -0.099109795  0.28037983 0.7481881
#> sample_16  -1.0371135 -0.081602027  0.39797105 0.7207444
#> sample_17  -1.1376894 -0.039077146  0.30469439 0.8720722
#> sample_18  -1.0254495 -0.178151689  0.47119287 0.7324084
#> sample_19  -0.5843153  0.162899121  0.12718104 0.2942351
#> sample_20  -0.8533964 -0.036635289  0.24521586 0.6448159
#> sample_21  -0.8289868  0.158399845 -0.10986414 0.7804511
#> sample_22  -1.0497562  0.003393703  0.38532831 0.6610342
#> sample_23  -0.5511553 -0.474194248  0.20026080 0.8250887
#> sample_24  -0.3152602 -0.592891931  0.23128351 0.6768686
#> sample_25  -1.9108918  0.329817881  0.59054414 0.9905298
#> sample_26  -0.9220226 -0.091674303  0.21505596 0.7986409
#> sample_27  -0.9986154 -0.339369739  0.55601431 0.7819708
#> sample_28  -0.9536749 -0.025688079  0.47579168 0.5035712
#> sample_29  -0.9168671 -0.253572849  0.45461221 0.7158277
#> sample_30  -1.0955650 -0.226527123  0.30977759 1.0123145
#> sample_31  -1.0421843  0.010965627  0.29281678 0.7384019
#> sample_32  -0.6336113 -0.335118354  0.12271474 0.8460150
#> sample_33  -0.6955800 -0.521226645  0.40303226 0.8137744
#> sample_34  -0.6789311 -0.053225244  0.03572224 0.6964341
#> sample_35  -1.0388106  0.014339364  0.26047243 0.7639988
#> sample_36  -1.0239895 -0.018467653  0.19978591 0.8426713
#> sample_37  -1.0454526 -0.243106120  0.10317012 1.1853886
#> sample_38  -0.6667645 -0.246910681  0.20507444 0.7086008
#> sample_39  -0.9550403 -0.527596259  0.57101603 0.9116205
#> sample_40  -0.6229518 -0.462609163  0.24888716 0.8366738
#> sample_41  -0.8025342  0.184852460 -0.25046561 0.8681473
#> sample_42  -0.9149728 -0.251678575  0.42620113 0.7404502
#> sample_43  -1.0514715  0.001678430  0.41486558 0.6349275
#> sample_44  -0.7155710 -0.464256616  0.47401302 0.7058146
#> sample_45  -0.6640675 -0.274602716 -0.09767201 1.0363422
#> sample_46  -0.5366452 -0.335974516  0.31065265 0.5619671
#> sample_47  -0.8538144 -0.551533574  0.51766483 0.8876832
#> sample_48  -1.1168508 -0.247812976  0.49258709 0.8720767
#> sample_49  -1.0239895 -0.018467653  0.19978591 0.8426713
#> sample_50  -0.8329769 -0.016215781  0.07273170 0.7764610
#> sample_51  -0.7891081 -0.366251298  0.27378606 0.8815734
#> sample_52  -0.9563038  0.017145358  0.38487014 0.5542883
#> sample_53  -0.8122442 -0.201335069  0.17514250 0.8384367
#> sample_54  -0.8743786  0.151474345  0.22423370 0.4986705
#> sample_55  -0.9365635 -0.158858958  0.54771124 0.5477112
#> sample_56  -1.1450395  0.140158717  0.14015872 0.8647221
#> sample_57  -0.7694043 -0.052726633  0.19340644 0.6287245
#> sample_58  -0.8527909 -0.133668231  0.49322956 0.4932296
#> sample_59  -0.2543142 -0.590786428  0.12216338 0.7229372
#> sample_60  -0.8068270 -0.383970120  0.45342667 0.7373704
#> sample_61  -0.6753115 -0.223326371 -0.10891602 1.0075539
#> sample_62  -0.6858224 -0.107744512  0.14897533 0.6445915
#> sample_63  -0.8852907 -0.422667189  0.51560245 0.7923555
#> sample_64  -0.6155306 -0.615530566  0.69162647 0.5394347
#> sample_65  -0.5151381 -0.689491451  0.34352355 0.8611060
#> sample_66  -0.6431269 -0.344633874  0.19167084 0.7960899
#> sample_67  -1.8796681  0.272094113  0.44271963 1.1648543
#> sample_68  -0.3347084 -0.167654357 -0.22029809 0.7226609
#> sample_69  -1.1314162  0.017206507  0.19545474 0.9187550
#> sample_70  -0.8007808 -0.311232600  0.26211338 0.8499000
#> sample_71  -0.8784970 -0.753333858  0.49099024 1.1408406
#> sample_72  -1.1364367 -0.146037987  0.47300122 0.8094735
#> sample_73  -0.6090761 -0.272603880  0.31518278 0.5664972
#> sample_74  -1.1860075 -0.448408593  0.52697106 1.1074451
#> sample_75  -0.7460073 -0.076957683  0.05961785 0.7633471
#> sample_76  -0.8998301 -0.236535924  0.27428970 0.8620764
#> sample_77  -1.1154626 -0.069494067  0.21140832 0.9735484
#> sample_78  -1.0084284 -0.105560653  0.21534707 0.8986420
#> sample_79  -0.9426299 -0.014643096  0.26720806 0.6900649
#> sample_80  -0.5813957 -0.303763916  0.13155416 0.7536054
#> sample_81  -0.7807485  0.145020963  0.06654935 0.5691782
#> sample_82  -1.0388988 -0.033376954  0.33058842 0.7416873
#> sample_83  -1.4590304  0.351078244  0.27557069 0.8323814
#> sample_84  -0.5651366 -0.099773302  0.18207785 0.4828320
#> sample_85  -0.6789311  0.035722242 -0.05322524 0.6964341
#> sample_86  -1.0565868  0.042025536  0.40975032 0.6048109
#> sample_87  -0.9356779 -0.007691104  0.20140069 0.7419683
#> sample_88  -0.4127800  0.085058378 -0.29141919 0.6191409
#> sample_89  -0.8254424 -0.158963512  0.20041049 0.7839955
#> sample_90  -1.3883715 -0.218300200  0.55753870 1.0491330
#> sample_91  -0.5352697 -0.275758496  0.17622663 0.6348016
#> sample_92  -0.6574894 -0.296476052  0.21434957 0.7396159
#> sample_93  -0.8273142 -0.010553025  0.03490935 0.8029578
#> sample_94  -0.6858224 -0.107744512  0.14897533 0.6445915
#> sample_95  -0.5858834 -0.370772006  0.27277823 0.6838772
#> sample_96  -0.6108392 -0.068514862  0.02043262 0.6589214
#> sample_97  -0.7643796 -0.197984173  0.45377779 0.5085860
#> sample_98  -1.0317657 -0.128897963  0.43457139 0.7260922
#> sample_99  -0.9848288 -0.081961115  0.06832109 0.9984689
#> sample_100 -0.5451050 -0.285593815  0.30219285 0.5285060