Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -0.9259636 -0.095615271  0.248156268 0.7734226
#> sample_2   -1.0421843  0.010965627  0.292816779 0.7384019
#> sample_3   -0.6394353 -0.407633721  0.268121716 0.7789473
#> sample_4   -1.0438019 -0.038279996  0.391282664 0.6907992
#> sample_5   -0.5907864 -0.254314192  0.122163380 0.7229372
#> sample_6   -0.5595322 -0.476150581 -0.031464759 1.0671475
#> sample_7   -1.2343283 -0.198236319  0.410827744 1.0217368
#> sample_8   -0.5660750 -0.350963590  0.106869504 0.8101691
#> sample_9   -1.2600285 -0.161416175  0.546768883 0.8746758
#> sample_10  -1.0451238  0.008026146  0.324363474 0.7127341
#> sample_11  -1.1720968  0.233245803  0.233245803 0.7056051
#> sample_12  -0.7407633 -0.174367806  0.146539914 0.7685912
#> sample_13  -0.7918246 -0.239756062  0.113883979 0.9176967
#> sample_14  -1.4452970  0.021040078  0.647937873 0.7763190
#> sample_15  -0.5479306 -0.037105001 -0.037105001 0.6221406
#> sample_16  -1.0144009 -0.225943573  0.451936135 0.7884084
#> sample_17  -1.2719095 -0.006243145  0.373246477 0.9049062
#> sample_18  -0.4577066 -0.968532205  0.528110213 0.8981286
#> sample_19  -1.4530962  0.199826863  0.392730529 0.8605388
#> sample_20  -0.7193005 -0.396527076  0.347050958 0.7687766
#> sample_21  -0.7694043 -0.052726633  0.193406436 0.6287245
#> sample_22  -1.2464786 -0.089025838  0.325918014 1.0095865
#> sample_23  -1.1449835  0.003639191  0.297400310 0.8439440
#> sample_24  -0.7023492 -0.076643289  0.273030460 0.5059620
#> sample_25  -0.7837943 -0.432396419  0.349304159 0.8668866
#> sample_26  -1.0092769 -0.220819568  0.393546735 0.8365498
#> sample_27  -0.2487912 -0.304361084 -0.053046656 0.6061990
#> sample_28  -1.1553480  0.171522936  0.171522936 0.8123021
#> sample_29  -0.6667645 -0.246910681  0.205074443 0.7086008
#> sample_30  -0.8689607  0.147973594 -0.268186804 0.9891739
#> sample_31  -1.1035201 -0.057551498  0.138193079 1.0228785
#> sample_32  -0.9158356  0.057613540  0.012151166 0.8460709
#> sample_33  -0.9204440 -0.257149814  0.536802066 0.6407918
#> sample_34  -1.0962069 -0.164648713  0.230664023 1.0301916
#> sample_35  -0.6724591 -0.094381278  0.042194257 0.7246461
#> sample_36  -0.7346478 -0.282662629  0.263881078 0.7534293
#> sample_37  -1.4462305  0.070116996  0.530932199 0.8451813
#> sample_38  -0.8327685 -0.221859410  0.365927255 0.6887006
#> sample_39  -0.3812944 -0.214240308  0.129531232 0.4660035
#> sample_40  -0.7217622 -0.210936593  0.083862947 0.8488359
#> sample_41  -0.9414877  0.031961460  0.195590884 0.7139353
#> sample_42  -0.8523864  0.173466538  0.009837114 0.6690827
#> sample_43  -0.8558325  0.006391012  0.207061708 0.6423798
#> sample_44  -0.3387031 -0.539373751  0.369482002 0.5085948
#> sample_45  -0.9956235 -0.269686470  0.373863769 0.8914462
#> sample_46  -0.6629392 -0.243085360  0.171858493 0.7341661
#> sample_47  -0.8159235 -0.326375251  0.414024814 0.7282739
#> sample_48  -1.2638221  0.001844309  0.308574576 0.9534032
#> sample_49  -0.5262416 -0.325570934  0.185254690 0.6665579
#> sample_50  -0.8254424 -0.158963512  0.200410489 0.7839955
#> sample_51  -0.6567537 -0.358260709  0.318625951 0.6963885
#> sample_52  -0.4977611 -0.058394453  0.026763355 0.5293922
#> sample_53  -0.3016164 -0.430828164 -0.084551927 0.8169965
#> sample_54  -1.5782800 -0.090202981  0.483143000 1.1853400
#> sample_55  -1.0509674  0.132802665  0.211274281 0.7068905
#> sample_56  -0.5383850 -0.278873801  0.208829405 0.6084294
#> sample_57  -0.4639232 -0.204412031 -0.263252531 0.9315878
#> sample_58  -0.5928919 -0.315260195  0.231283512 0.6768686
#> sample_59  -0.7667936 -0.415395712  0.220593055 0.9615963
#> sample_60  -0.4813618 -0.293309561  0.228879821 0.5457915
#> sample_61  -1.2464786 -0.089025838  0.325918014 1.0095865
#> sample_62  -0.6055237 -0.063199369 -0.017736994 0.6864600
#> sample_63  -0.6985074 -0.120429541  0.309133119 0.5098038
#> sample_64  -1.2600285 -0.161416175  0.546768883 0.8746758
#> sample_65  -0.7449947 -0.293009570  0.384870139 0.6531341
#> sample_66  -0.6751012 -0.199677508  0.196737764 0.6780409
#> sample_67  -1.2498444  0.158922824  0.065832401 1.0250892
#> sample_68  -0.3455456 -0.560656962  0.070614815 0.8355877
#> sample_69  -0.9072974 -0.076949131  0.109636825 0.8746097
#> sample_70  -1.3024967  0.149755603  0.410481865 0.7422593
#> sample_71  -0.7525950 -0.035917296  0.053030190 0.7354821
#> sample_72  -0.9439768 -0.063618120  0.365944540 0.6416504
#> sample_73  -1.1384572 -0.148058541  0.501286017 0.7852298
#> sample_74  -0.9414877  0.031961460  0.195590884 0.7139353
#> sample_75  -1.0573904  0.126379704  0.277610674 0.6534000
#> sample_76  -1.1536765  0.088036676  0.251666100 0.8139737
#> sample_77  -1.1001082 -0.231070320  0.342275661 0.9889028
#> sample_78  -0.6535484 -0.292535084  0.181249268 0.7648342
#> sample_79  -0.8493218  0.056386790  0.098059486 0.6948756
#> sample_80  -0.7965207 -0.306972449  0.229332260 0.8741609
#> sample_81  -0.5401479 -0.339477200  0.398121743 0.4815034
#> sample_82  -1.1424500 -0.152051305  0.583060074 0.7114412
#> sample_83  -1.2585799 -0.101127096  0.421062287 0.9386447
#> sample_84  -2.2296291 -0.032404523  1.015564033 1.2464696
#> sample_85  -0.8467029 -0.029941729  0.179150069 0.6974945
#> sample_86  -0.8385397 -0.227630631  0.479701185 0.5864692
#> sample_87  -0.8998301 -0.236535924  0.274289700 0.8620764
#> sample_88  -0.5699852  0.066003578  0.025998243 0.4779834
#> sample_89  -0.9998782 -0.340632568  0.582531043 0.7579797
#> sample_90  -0.7713934 -0.580338141  0.601655757 0.7500758
#> sample_91  -0.7020128 -0.527659459  0.487571220 0.7421011
#> sample_92  -0.9330312  0.241088660 -0.052672459 0.7446150
#> sample_93  -0.5472486 -0.179523774  0.127206493 0.5995658
#> sample_94  -0.9553295  0.143282758  0.181749039 0.6302977
#> sample_95  -1.0476437  0.005506256  0.355180005 0.6869574
#> sample_96  -0.6314789 -0.089154586  0.227182742 0.4934507
#> sample_97  -0.8048064 -0.381949566  0.425141874 0.7616141
#> sample_98  -0.8800333 -0.279259428  0.218579000 0.9407137
#> sample_99  -0.6574894 -0.296476052  0.214349571 0.7396159
#> sample_100 -0.8868783 -0.424254748  0.542588263 0.7685448