Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

Usage

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2       cat_3     cat_4
#> sample_1   -0.3941013 -0.217170606 -0.11451645 0.7257884
#> sample_2   -0.7518917 -0.241066059  0.34672061 0.6462371
#> sample_3   -1.4456607  0.207262356  0.32740667 0.9109916
#> sample_4   -1.0077032 -0.002181370  0.09090905 0.9189756
#> sample_5   -0.7684574 -0.149418155  0.36140747 0.5564681
#> sample_6   -0.5149865  0.009538001  0.12505089 0.3803976
#> sample_7   -0.6856954 -0.157627950  0.22186167 0.6214617
#> sample_8   -1.2577715 -0.159159187  0.51872052 0.8982101
#> sample_9   -0.6019200 -0.265447757  0.22225545 0.6451123
#> sample_10  -0.7112208 -0.042171125 -0.14482528 0.8982172
#> sample_11  -0.9404275 -0.512983458  0.42905977 1.0243512
#> sample_12  -0.5095981 -0.115693843  0.29749331 0.3277987
#> sample_13  -0.4669653 -0.337753612  0.17307201 0.6316469
#> sample_14  -0.4924133 -0.248791233  0.28342558 0.4577790
#> sample_15  -1.1684474  0.073265694  0.40973793 0.6854438
#> sample_16  -1.2451505 -0.146538211  0.40000550 0.9916832
#> sample_17  -1.4481248  0.559343274  0.06822272 0.8205588
#> sample_18  -0.7694043  0.193406436 -0.05272663 0.6287245
#> sample_19  -0.4829708 -0.043604149 -0.08906652 0.6156415
#> sample_20  -0.9217322 -0.144027592  0.28810576 0.7776540
#> sample_21  -1.0306880  0.022461905  0.19308742 0.8151387
#> sample_22  -0.8467029 -0.029941729  0.17915007 0.6974945
#> sample_23  -1.3066476  0.302790278  0.26574901 0.7381083
#> sample_24  -0.9774852 -0.389698545  0.45759932 0.9095844
#> sample_25  -0.4590519 -0.270999685  0.02379986 0.7062517
#> sample_26  -0.8693634  0.036345247  0.32933237 0.5036858
#> sample_27  -1.1682093  0.028041423  0.52952118 0.6106467
#> sample_28  -0.8080212 -0.318472925  0.32507731 0.8014168
#> sample_29  -1.0584622  0.165313226  0.24082078 0.6523282
#> sample_30  -0.9325239 -0.102175555  0.31180024 0.7228992
#> sample_31  -0.8136661 -0.324117900  0.38502962 0.7527544
#> sample_32  -0.8691787  0.156674249  0.15667425 0.5558302
#> sample_33  -0.8136661 -0.324117900  0.38502962 0.7527544
#> sample_34  -1.6351481  0.132513824  0.52981562 0.9728186
#> sample_35  -0.5545296 -0.745584825  0.31730938 0.9828050
#> sample_36  -0.7749049 -0.058227172  0.25811016 0.5750219
#> sample_37  -0.8582451  0.129141591  0.08913626 0.6399672
#> sample_38  -0.5691836 -0.014872849  0.10527146 0.4787850
#> sample_39  -1.0996105 -0.297264062  0.44631397 0.9505606
#> sample_40  -0.6769802 -0.586008376  0.38937127 0.8736173
#> sample_41  -0.6847410  0.029912432 -0.01357268 0.6684012
#> sample_42  -0.8136661 -0.324117900  0.38502962 0.7527544
#> sample_43  -0.6167183 -0.224676166  0.30754065 0.5338538
#> sample_44  -0.9094702  0.018516593  0.01851659 0.8724370
#> sample_45  -0.7728882 -0.498451305  0.39354673 0.8777927
#> sample_46  -0.9164612 -0.194326524  0.32786286 0.7829249
#> sample_47  -0.5049295 -0.433470514  0.20656684 0.7318331
#> sample_48  -0.8788664  0.219745920  0.21974592 0.4393745
#> sample_49  -0.9013283 -0.070979971  0.07212087 0.9001874
#> sample_50  -0.8113087 -0.259240115  0.25158551 0.8189633
#> sample_51  -0.7372008 -0.737200814  0.73720081 0.7372008
#> sample_52  -0.4499867 -0.153720879 -0.15372088 0.7574285
#> sample_53  -0.5391269 -0.171402149  0.05685650 0.6536726
#> sample_54  -1.0239895 -0.018467653  0.19978591 0.8426713
#> sample_55  -0.8899540 -0.167819278  0.12698026 0.9307930
#> sample_56  -0.5715798 -0.423159774  0.25259566 0.7421439
#> sample_57  -0.5560889  0.079899827 -0.09072569 0.5669148
#> sample_58  -0.7807485  0.145020963  0.06654935 0.5691782
#> sample_59  -1.0172029 -0.114335157  0.28208012 0.8494579
#> sample_60  -0.6456366 -0.413835024  0.32974301 0.7297287
#> sample_61  -1.4378926  0.126082902  0.37221597 0.9395938
#> sample_62  -0.8473036 -0.078170541  0.25130866 0.6741655
#> sample_63  -0.7317714 -0.631687942  0.36684089 0.9966185
#> sample_64  -0.7400818 -0.350617029  0.50585521 0.5848436
#> sample_65  -0.8380757 -0.021314606  0.10930558 0.7500848
#> sample_66  -0.6093503 -0.114654073  0.06359416 0.6604102
#> sample_67  -0.6043535 -0.520971911  0.23044418 0.8948813
#> sample_68  -0.7320793 -0.221253716  0.15522386 0.7981092
#> sample_69  -0.8799610 -0.216666774  0.13697327 0.9596545
#> sample_70  -1.4608861  0.148551808  0.57599582 0.7363385
#> sample_71  -0.5050446 -0.158768351  0.30204685 0.3617661
#> sample_72  -0.6267791 -0.466436466  0.28077794 0.8124376
#> sample_73  -0.8025342 -0.250465611  0.18485246 0.8681473
#> sample_74  -0.6609839 -0.299970576  0.24657313 0.7143814
#> sample_75  -0.6388233 -0.277809913  0.07583013 0.8408030
#> sample_76  -0.9193604 -0.197225633  0.35830017 0.7582858
#> sample_77  -0.7633115 -0.094261885  0.19949923 0.6580742
#> sample_78  -0.6985828  0.016070587  0.09774862 0.5847636
#> sample_79  -1.5964411 -0.108364081  0.63203598 1.0727692
#> sample_80  -0.5399018 -0.224820790  0.13455321 0.6301694
#> sample_81  -1.0544135  0.129356563  0.24486945 0.6801875
#> sample_82  -1.5783170 -0.156931339  0.58664669 1.1486017
#> sample_83  -0.6044879 -0.109791683  0.02678385 0.6874958
#> sample_84  -0.8046081 -0.315059903  0.29400416 0.8256639
#> sample_85  -1.0358331 -0.030311194  0.29916801 0.7669762
#> sample_86  -0.9193604 -0.197225633  0.35830017 0.7582858
#> sample_87  -0.9563038  0.017145358  0.38487014 0.5542883
#> sample_88  -0.4695778 -0.546538885  0.16641092 0.8497058
#> sample_89  -0.5753821 -0.360270754  0.17603395 0.7596189
#> sample_90  -0.9292059  0.207872709 -0.04884714 0.7701803
#> sample_91  -0.7575782 -0.040900474  0.08971971 0.7087589
#> sample_92  -1.0905782 -0.221540315  0.27629811 1.0358204
#> sample_93  -1.0419917 -0.086480230  0.48406463 0.6444073
#> sample_94  -1.0497562  0.003393703  0.38532831 0.6610342
#> sample_95  -1.0239895 -0.018467653  0.19978591 0.8426713
#> sample_96  -0.7330477 -0.343582980  0.36556454 0.7110662
#> sample_97  -0.7527864 -0.133747230  0.17298304 0.7135506
#> sample_98  -0.9407658 -0.110417545  0.43071343 0.6204700
#> sample_99  -0.5715798 -0.423159774  0.25259566 0.7421439
#> sample_100 -0.4187169 -0.356196559  0.25816974 0.5167437