Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3     cat_4
#> sample_1   -0.7837943 -0.432396419  0.349304159 0.8668866
#> sample_2   -0.9848288 -0.081961115  0.068321088 0.9984689
#> sample_3   -1.6377441  0.223008289  0.423678985 0.9910568
#> sample_4   -0.6936642 -0.022495909  0.102667234 0.6134929
#> sample_5   -0.4111520 -0.411152042  0.297995479 0.5243086
#> sample_6   -0.8768811 -0.491218582  0.659353445 0.7087462
#> sample_7   -0.5080248 -0.090289607 -0.192943761 0.7912582
#> sample_8   -1.0941732 -0.103774466  0.147539962 1.0504077
#> sample_9   -1.0514715  0.001678430  0.414865585 0.6349275
#> sample_10  -0.7305383 -0.341073549  0.335813111 0.7357988
#> sample_11  -0.6218140 -0.177128151  0.270184067 0.5287581
#> sample_12  -0.8607511  0.044957508  0.202143092 0.6136505
#> sample_13  -0.9436368 -0.432811210  0.318604878 1.0578432
#> sample_14  -1.0550239  0.043588347  0.380060583 0.6313750
#> sample_15  -0.3669019 -0.308061429  0.143923695 0.5310397
#> sample_16  -0.9320791 -0.154374533  0.409094824 0.6773588
#> sample_17  -1.4153266 -0.001633253  0.394782019 1.0221778
#> sample_18  -1.4229846 -0.064861139  0.554178070 0.9336677
#> sample_19  -0.4798506  0.044673825 -0.183584827 0.6187616
#> sample_20  -0.5677451 -0.490784081  0.356513780 0.7020154
#> sample_21  -0.7615468 -0.195151281  0.368318076 0.5883800
#> sample_22  -0.8747491 -0.273975216  0.183857878 0.9648664
#> sample_23  -0.5958314 -0.318199676  0.262830207 0.6512009
#> sample_24  -0.7564910 -0.137451764  0.206319776 0.6876230
#> sample_25  -0.6789311 -0.053225244  0.035722242 0.6964341
#> sample_26  -0.7067551 -0.455440649  0.359596349 0.8025994
#> sample_27  -0.4820733 -0.559034388  0.265141055 0.7759667
#> sample_28  -1.2415739 -0.205482010  0.471404650 0.9756513
#> sample_29  -1.0523829 -0.481838009  0.460205219 1.0740157
#> sample_30  -0.4071009 -0.407100911  0.236449329 0.5777525
#> sample_31  -0.9100852 -0.246791004  0.367575298 0.7893009
#> sample_32  -0.6357143 -0.403912722  0.236124633 0.8035024
#> sample_33  -0.4723718 -0.228749690  0.052152695 0.6489688
#> sample_34  -0.7449947 -0.293009570  0.384870139 0.6531341
#> sample_35  -1.1198367  0.121876424  0.028786001 0.9691743
#> sample_36  -0.9439768 -0.063618120  0.365944540 0.6416504
#> sample_37  -1.0204143 -0.231956891  0.561994988 0.6903762
#> sample_38  -1.1830324  0.183843870  0.456710857 0.5424777
#> sample_39  -0.9016379 -0.367555390  0.634596633 0.6345966
#> sample_40  -0.7873492  0.099953963  0.175461516 0.5119338
#> sample_41  -0.3977385 -0.397738469  0.138566240 0.6569107
#> sample_42  -0.6113335 -0.023546804 -0.023546804 0.6584271
#> sample_43  -1.0242261 -0.121358350  0.345261181 0.8003232
#> sample_44  -0.5856929 -0.308061137  0.165723215 0.7280308
#> sample_45  -0.9578971  0.015552045  0.442996060 0.4993490
#> sample_46  -0.6567537 -0.358260709  0.318625951 0.6963885
#> sample_47  -0.5005224 -0.185441323 -0.132797589 0.8187613
#> sample_48  -0.3223767 -0.711841510  0.175461685 0.8587566
#> sample_49  -0.6422817 -0.742365186  0.456330561 0.9283164
#> sample_50  -0.4351304 -0.583550396  0.402266399 0.6164144
#> sample_51  -0.5730391 -0.496078085  0.442191554 0.6269257
#> sample_52  -1.0471408 -0.041618944  0.449501609 0.6392581
#> sample_53  -0.5114898 -0.440030797  0.270210816 0.6813097
#> sample_54  -0.9414877  0.031961460  0.195590884 0.7139353
#> sample_55  -0.8892002 -0.288426358  0.284919623 0.8927070
#> sample_56  -0.6722917 -0.311278326  0.396906732 0.5866633
#> sample_57  -0.6165227 -0.028736019  0.014749093 0.6305096
#> sample_58  -0.7303348 -0.407561448  0.568948144 0.5689481
#> sample_59  -0.6818427 -0.507489284  0.280968076 0.9083639
#> sample_60  -1.1364367 -0.146037987  0.473001221 0.8094735
#> sample_61  -0.7918062 -0.302257997  0.195580432 0.8984838
#> sample_62  -0.6922453 -0.114167473  0.215311728 0.5911011
#> sample_63  -0.7929530  0.273398411  0.094350180 0.4252044
#> sample_64  -0.6373225  0.114093549  0.035621934 0.4876071
#> sample_65  -0.8403516 -0.071218550  0.185501297 0.7260689
#> sample_66  -0.4669653 -0.337753612  0.173072011 0.6316469
#> sample_67  -1.2671623 -0.109709463  0.509329746 0.8675420
#> sample_68  -0.5342645 -0.166539759  0.020046197 0.6807581
#> sample_69  -1.1142462 -0.245208362  0.463939160 0.8955154
#> sample_70  -0.7128727 -0.390099299  0.285656139 0.8173158
#> sample_71  -0.3993064 -0.336786081  0.058526655 0.6775659
#> sample_72  -0.9632451 -0.452419492  0.471839409 0.9438252
#> sample_73  -0.8071506 -0.255082015  0.218702337 0.8435303
#> sample_74  -0.4439984 -0.314786653 -0.004631725 0.7634168
#> sample_75  -0.7923767 -0.440978786  0.437571618 0.7957838
#> sample_76  -1.2482131  0.017453307  0.204039263 1.0267205
#> sample_77  -0.4034581 -0.470149439  0.367247350 0.5063602
#> sample_78  -1.2262770 -0.256876394  0.453365220 1.0297881
#> sample_79  -0.9565558 -0.445730130  0.412931489 0.9893544
#> sample_80  -1.2419715  0.023694823  0.166795667 1.0514811
#> sample_81  -0.9070705 -0.243776249  0.337253634 0.8135931
#> sample_82  -1.3089296  0.184995435  0.436309863 0.6876243
#> sample_83  -1.1313127 -0.140913982  0.414611820 0.8576148
#> sample_84  -0.7467273 -0.472290501  0.200653972 1.0183639
#> sample_85  -0.6191178 -0.227075705  0.366699002 0.4794945
#> sample_86  -1.0509674  0.132802665  0.211274281 0.7068905
#> sample_87  -0.6361647  0.152292677 -0.004892907 0.4887649
#> sample_88  -1.0147081  0.038441785  0.083904159 0.8923622
#> sample_89  -0.9294581 -0.099109795  0.280379827 0.7481881
#> sample_90  -0.6789311 -0.053225244  0.035722242 0.6964341
#> sample_91  -1.2680822  0.341355697 -0.002415842 0.9291424
#> sample_92  -1.6278540  0.089797487  0.537109705 1.0009468
#> sample_93  -0.6361986 -0.216344784 -0.010492730 0.8630361
#> sample_94  -0.7221233  0.185433718  0.219919895 0.3167697
#> sample_95  -0.5359044 -0.619285983  0.356093665 0.7990967
#> sample_96  -0.7522596 -0.185864084  0.246269271 0.6918544
#> sample_97  -0.9618235  0.055110770  0.439069673 0.4676430
#> sample_98  -0.5958314 -0.318199676  0.262830207 0.6512009
#> sample_99  -1.0835219 -0.281175416  0.321820668 1.0428766
#> sample_100 -0.9994380 -0.152140159  0.224337412 0.9272408