Centered log ratio transform

Compute the centered log ratio (CLR) transform of a count matrix.

clr(counts, pseudocount = 0.5)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

Value

matrix of CLR transformed counts

Details

The CLR of a vector x of counts in D categories is defined as clr(x) = log(x) - mean(log(x)). For details see van den Boogaart and Tolosana-Delgado (2013).

References

Van den Boogaart KG, Tolosana-Delgado R (2013). Analyzing compositional data with R, volume 122. Springer. https://link.springer.com/book/10.1007/978-3-642-36809-7.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# centered log ratio
clr(counts)
#>                 cat_1        cat_2        cat_3        cat_4
#> sample_1   -0.8699616  0.077419762  0.263136908  0.529404886
#> sample_2   -0.9070705 -0.243776249  0.337253634  0.813593081
#> sample_3   -0.6730777 -0.253223899  0.268965484  0.657336159
#> sample_4   -0.6229518 -0.462609163  0.248887156  0.836673821
#> sample_5   -0.5224805 -0.384330204  0.259220036  0.647590710
#> sample_6   -0.7330477 -0.343582980  0.365564542  0.711066185
#> sample_7   -1.2088453 -0.172753347  0.243407050  1.138191577
#> sample_8   -0.6870676 -0.211643900  0.546886040  0.351825457
#> sample_9   -0.8659540  0.039754632  0.267144475  0.559054883
#> sample_10  -0.8955700 -0.232275773  0.241508580  0.886337183
#> sample_11  -1.0898110 -0.358923501  0.488374360  0.960360150
#> sample_12  -1.6278540  0.089797487  0.537109705  1.000946819
#> sample_13  -0.8473036 -0.078170541  0.251308660  0.674165511
#> sample_14  -0.7658348 -0.491397940  0.332777503  0.924455223
#> sample_15  -0.6167183 -0.224676166  0.307540647  0.533853773
#> sample_16  -0.6794255 -0.259571614  0.359467594  0.579529479
#> sample_17  -0.8332840 -0.114161362  0.192568906  0.754876485
#> sample_18  -0.5059728  0.058557011 -0.023121020  0.470536800
#> sample_19  -1.4530962  0.199826863  0.392730529  0.860538768
#> sample_20  -0.5857113 -0.043387026 -0.141025495  0.770123837
#> sample_21  -0.8677506  0.079630699  0.230861668  0.557258253
#> sample_22  -0.5432901 -0.283778925  0.271746878  0.555322168
#> sample_23  -0.6945341  0.020119259  0.061791955  0.612622913
#> sample_24  -1.1259001  0.159298098  0.022722563  0.943879485
#> sample_25  -1.2507545  0.283175841 -0.037731879  1.005310558
#> sample_26  -0.5967958 -0.596795838  0.310761214  0.882830463
#> sample_27  -0.7206160 -0.331151247  0.242194734  0.809572527
#> sample_28  -1.0122399 -0.164942066  0.322761141  0.854420851
#> sample_29  -0.8376467 -0.226737613  0.452020830  0.612363480
#> sample_30  -0.6892685 -0.111190614  0.182570504  0.617888575
#> sample_31  -1.1149578 -0.183399614  0.363144092  0.935213341
#> sample_32  -0.7910005 -0.124521536 -0.021867382  0.937389386
#> sample_33  -0.9176146 -0.087266254  0.180997732  0.823883078
#> sample_34  -0.6357143 -0.403912722  0.236124633  0.803502426
#> sample_35  -0.6149034 -0.222861276  0.277094675  0.560669966
#> sample_36  -1.2917561  0.071548710  0.421222459  0.798984964
#> sample_37  -0.3701807 -0.793037503  0.305574786  0.857643368
#> sample_38  -1.1733376  0.153533371  0.339250517  0.680553681
#> sample_39  -0.7488858 -0.182490348  0.213924925  0.717451246
#> sample_40  -1.0358331 -0.030311194  0.299168007  0.766976246
#> sample_41  -1.0081449 -0.282207865  0.517911435  0.772441300
#> sample_42  -1.1648950  0.031355778  0.444542932  0.688996270
#> sample_43  -0.8615190 -0.044757831  0.368429323  0.537847475
#> sample_44  -1.0898110 -0.358923501  0.488374360  0.960360150
#> sample_45  -0.6520393 -0.420237650  0.417159139  0.655117776
#> sample_46  -0.9450659  0.028383205  0.229053901  0.687628834
#> sample_47  -1.1392705  0.102442665  0.145927777  0.890900025
#> sample_48  -0.9276970 -0.149992460  0.349963492  0.727725997
#> sample_49  -1.1675534  0.159317558  0.274830445  0.733405379
#> sample_50  -0.6609839 -0.299970576  0.246573130  0.714381369
#> sample_51  -0.8007808 -0.311232600  0.262113380  0.849900045
#> sample_52  -0.8510963 -0.131973628  0.409157345  0.573912578
#> sample_53  -0.7306545 -0.164259038  0.074970651  0.819942900
#> sample_54  -1.0287655 -0.073254006  0.306235616  0.795783841
#> sample_55  -0.6211266 -0.033339978  0.051817831  0.602648789
#> sample_56  -0.6629392 -0.243085360  0.171858493  0.734166072
#> sample_57  -0.7378343 -0.285849176  0.295180707  0.728502769
#> sample_58  -1.1042298 -0.235192000  0.373872063  0.965549784
#> sample_59  -0.8327685  0.365927255 -0.221859410  0.688700647
#> sample_60  -1.0421705  0.181604897  0.099926866  0.760638771
#> sample_61  -1.2680822 -0.002415842  0.341355697  0.929142362
#> sample_62  -0.7598315 -0.193436008  0.338780805  0.614486687
#> sample_63  -0.9085904 -0.307816588  0.492302712  0.724104326
#> sample_64  -0.2933096 -0.481361793  0.228879821  0.545791532
#> sample_65  -0.5299653 -0.112230069 -0.064602020  0.706797358
#> sample_66  -1.2549585 -0.097505718  0.390197488  0.962266737
#> sample_67  -0.8504265 -0.316344041  0.123022619  1.043747948
#> sample_68  -0.9217322 -0.144027592  0.288105763  0.777653989
#> sample_69  -0.5358293 -0.458868295  0.095442441  0.899255189
#> sample_70  -0.5949680 -0.258495765  0.156448087  0.697015680
#> sample_71  -0.9314963 -0.003509531  0.167115986  0.767889847
#> sample_72  -0.6333150  0.118101115  0.517257054 -0.002043196
#> sample_73  -0.6036491 -0.603649103  0.371730546  0.835567660
#> sample_74  -0.5983513 -0.320719565  0.293646737  0.625424130
#> sample_75  -1.3449743 -0.497676478  0.852250239  0.990400577
#> sample_76  -0.7276392 -0.338174440  0.305375799  0.760437848
#> sample_77  -0.8737195 -0.210425258  0.099729670  0.984415063
#> sample_78  -1.0619809  0.237302072  0.200260800  0.624418041
#> sample_79  -1.2781293  0.085175546  0.294267343  0.898686408
#> sample_80  -1.2713245  0.044352291  0.301072138  0.925900074
#> sample_81  -0.8110359  0.355398979 -0.321487681  0.777124608
#> sample_82  -0.3377536 -0.466965344  0.173072011  0.631646945
#> sample_83  -0.3750784 -0.208024353  0.060239634  0.522863156
#> sample_84  -0.8183319 -0.266263308  0.314766574  0.769828624
#> sample_85  -1.0210736  0.077538732  0.077538732  0.865996092
#> sample_86  -0.7749049 -0.058227172  0.258110156  0.575021867
#> sample_87  -1.1466676 -0.100699089  0.493075617  0.754291117
#> sample_88  -0.8258715 -0.273802921  0.404076788  0.695597636
#> sample_89  -0.6431269 -0.344633874  0.191670835  0.796089900
#> sample_90  -0.5930054 -0.050681109 -0.098309158  0.741995667
#> sample_91  -0.3963807 -0.463072104  0.247169510  0.612283323
#> sample_92  -0.7837943 -0.432396419  0.349304159  0.866886565
#> sample_93  -0.8814569 -0.347374410  0.328381028  0.900450277
#> sample_94  -1.2695612  0.182691131  0.139206019  0.947664047
#> sample_95  -0.9393824 -0.011395638  0.234737432  0.716040616
#> sample_96  -1.1599143  0.166956599  0.206961934  0.785995808
#> sample_97  -0.9478186 -0.019831867  0.329841881  0.637808625
#> sample_98  -0.6431269 -0.344633874  0.191670835  0.796089900
#> sample_99  -0.7429958 -0.291010650  0.355616515  0.678389908
#> sample_100 -0.5579063 -0.003595568 -0.003595568  0.565097438