Count ratio uncertainty modeling based linear regression

Count ratio uncertainty modeling based linear regression (crumblr) returns CLR-transformed counts and observation-level inverse-variance weights for use in weighted linear models.

crumblr(counts, pseudocount = 0.5, tau = NULL)

# S4 method for matrix
crumblr(counts, pseudocount = 0.5, tau = NULL)

# S4 method for data.frame
crumblr(counts, pseudocount = 0.5, tau = NULL)

Arguments

counts

count data with samples as rows and variables are columns

pseudocount

added to counts to avoid issues with zeros

tau

overdispersion parameter for Dirichlet multinomial. If NULL, estimate from observed counts.

Value

An EList object with the following components:

E: numeric matrix of CLR transformed counts

weights: numeric matrix of observation-level inverse-variance weights

Details

Evalute the centered log ratio (CLR) transform of the count matrix, and the asymptotic theoretical variances of each transformed observation. The asymptotic normal approximation is increasingly accurate for small overdispersion \(\tau\), large total counts \(C\), and large proportions \(p\), but shows good agreement with the empirical results most situtations. In practice, it is often reasonable to assume a sufficient number of counts before a variable is included in an analysis anyway. But the feasability of this assumption is up to the user to determine.

See also

Examples

# set probability of each category
prob = c(0.1, 0.2, 0.3, 0.5)

# number of total counts
countsTotal = 300

# number of samples
n_samples = 100

# simulate info for each sample
info = data.frame(Age = rgamma(n_samples, 50, 1))
rownames(info) = paste0("sample_", 1:n_samples)

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = n_samples, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# run crumblr on counts
cobj = crumblr(counts)

# run standard limma analysis on crumblr results
library(limma)

design = model.matrix(~Age, info)
fit = lmFit(cobj, design)
fit = eBayes(fit)

topTable(fit, coef="Age", sort.by="none")
#>              logFC    AveExpr          t   P.Value adj.P.Val         B
#> cat_1  0.002159459 -0.9285324  0.5375914 0.5911642 0.7047478 -7.666312
#> cat_2 -0.002381596 -0.1405127 -0.7883882 0.4309462 0.7047478 -7.784716
#> cat_3  0.001003102  0.2741961  0.3791958 0.7047478 0.7047478 -8.155996
#> cat_4 -0.001528041  0.7948490 -0.6511922 0.5153040 0.7047478 -8.135720